### 2023-05-01 #5427 @editorialbot
[PRE REVIEW]: baseflow: a MATLAB and GNU Octave package for baseflow recession analysis
**Submitting author:** @mgcooper (<a href="http://orcid.org/0000-0002-0165-209X">Matthew Cooper</a>)
**Repository:** https://github.com/mgcooper/baseflow
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
