### 2023-05-01 #5427 @kthyng
Hi @mgcooper and thanks for your submission! A few notes:

1. Your package looks like it's in decent shape — but I don't see [tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) offhand.
2. Also how does a user easily see the docs? I see the files are present.
3. We usually have a hard time finding matlab reviewers so please go through the reviewer list carefully to find suggestions, and if you have other suggestions that aren't conflicted with you, please include them. Just don't include "@" beforehand so they are not pinged here.
4. You mentioned a paper that was already published with this software. Can you provide a summary of what was covered in that paper so that it is clear this review would be on a distinct scholarly effort?
