### 2023-05-02 #5427 @mgcooper
Hi @kthyng , thank you for the questions. Here are my answers:

1. The tests are located in +bfra/+test, they can be run in a matlab session by typing `bfra.test.runtests` at the command window and pressing enter. The tests run with the github workflow in .github/workflows/ci.yml, configured to run on push to dev, but I confirmed all tests pass on joss branch. I added a note to the joss branch README for users who want to run the tests locally.
2. The documentation is configured to display in the matlab help browser. I have a landing page published by github pages on the main branch that provides a basic introduction to the toolbox (link in the README), but I removed the link from the joss branch README because I get confused when the same thing exists in two places (the landing page on github pages is identical to the one displayed in matlab's help browser, and I think this might confuse users). That said, I am working right now on getting the docs integrated into the repo so they display identically to the matlab help browser. Until then, you can click on the html files in the docs/html folder to verify them, and if the paper is sent for review the referees can use the matlab help browser while I get the docs configured on the repo.
3. Based on the google sheet filtered by Matlab and by discipline, I generated a list of referees who appear as good choices included at the bottom of this message.
4. The paper (https://onlinelibrary.wiley.com/doi/abs/10.1029/2022WR033154) presented a theoretical derivation of the relationship between the rate at which streamflow drains a river basin, and the amount of groundwater stored upstream in river basin soils. The software was developed to estimate the parameters in the equation (the software was not the topic of the paper). An important theoretical detail revealed in the paper is that the "lifetime" of water in a river draining a river basin follows a Pareto distribution. One key feature of the software is the estimation of the parameters of the Pareto distribution, which describes how long a river would flow before it went dry, if it never rained or snowed again. The specific topic addressed by the paper was how water stored in soils is changing over time in permafrost-dominated river basins where permafrost (frozen soil) is melting, which creates more space for water storage. The effect of additional storage can be detected in streamflow with careful signal processing. The parameters estimated by the software are used in many other areas of hydrology in addition to the specific topic addressed in that paper. Please note that this information is mentioned in the README and was briefly described in the submission note to the editor, and is the topic of the demo bfra_kuparuk.mlx, and expanded upon in the other demos, including the theoretical basis demos and help pages, as well as the paper draft, if more information is needed please reach out or refer to those resources.

Many thanks!

List of suitable referees:

amabdallah
andres-patrignani
chadagreene
mvdh7
fgnievinski
ZijieZhaoMMHW
chouj
alberto-battistel
ShiyongYu-cloud
WalBytes
ajmalkoti

If referees are not found, please feel free to reach out and I will try to provide a list of candidates.

Matt
