+++ paper.bib
@@ -9,6 +9,7 @@
   pages = {87--92},
   issn = {0098-3004},
   doi = {10.1016/j.cageo.2016.10.005},
+  urldate = {2023-02-14},
   langid = {english}
 }
 
@@ -17,6 +18,7 @@
   author = {{Arciniega-Esparza}, Sa{\'u}l},
   year = {2018},
   address = {{Natick, Massachusetts}},
+  urldate = {2023-02-11},
   howpublished = {MATLAB Central File Exchange, The MathWorks Inc.}
 }
 
@@ -31,6 +33,7 @@
   pages = {637--643},
   issn = {1944-7973},
   doi = {10.1029/WR013i003p00637},
+  urldate = {2017-04-17},
   langid = {english}
 }
 
@@ -45,6 +48,7 @@
   pages = {233--240},
   issn = {1944-7973},
   doi = {10.1029/97WR03068},
+  urldate = {2017-05-03},
   langid = {english}
 }
 
@@ -59,7 +63,8 @@
   pages = {661--703},
   publisher = {{Society for Industrial and Applied Mathematics}},
   issn = {0036-1445},
-  doi = {10.1137/070710111}
+  doi = {10.1137/070710111},
+  urldate = {2021-05-18}
 }
 
 @article{cooper_2018_WRR,
@@ -72,6 +77,7 @@
   pages = {5602--5619},
   issn = {1944-7973},
   doi = {10.1029/2018WR022816},
+  urldate = {2019-03-05},
   copyright = {\textcopyright 2018. American Geophysical Union. All Rights Reserved.},
   langid = {english}
 }
@@ -86,6 +92,7 @@
   pages = {e2022WR033154},
   issn = {1944-7973},
   doi = {10.1029/2022WR033154},
+  urldate = {2023-01-17},
   langid = {english}
 }
 
@@ -102,6 +109,7 @@
   publisher = {{Copernicus GmbH}},
   issn = {1027-5606},
   doi = {10.5194/hess-21-65-2017},
+  urldate = {2021-04-27},
   langid = {english}
 }
 
@@ -116,6 +124,7 @@
   pages = {104983},
   issn = {1364-8152},
   doi = {10.1016/j.envsoft.2021.104983},
+  urldate = {2021-06-14},
   langid = {english}
 }
 
@@ -130,6 +139,7 @@
   pages = {973--983},
   issn = {1944-7973},
   doi = {10.1029/WR004i005p00973},
+  urldate = {2016-10-27},
   langid = {english}
 }
 
@@ -143,7 +153,8 @@
   number = {9},
   publisher = {{John Wiley \& Sons, Ltd}},
   issn = {0043-1397},
-  doi = {10.1029/2008WR007392}
+  doi = {10.1029/2008WR007392},
+  urldate = {2021-08-25}
 }
 
 @article{jachens_2020_HESS,
@@ -170,7 +181,8 @@
   number = {2},
   publisher = {{John Wiley \& Sons, Ltd}},
   issn = {0043-1397},
-  doi = {10.1029/2008WR006912}
+  doi = {10.1029/2008WR006912},
+  urldate = {2021-03-10}
 }
 
 @misc{matlab_2020_,
@@ -191,6 +203,7 @@
   pages = {29--43},
   issn = {0309-1708},
   doi = {10.1016/j.advwatres.2017.07.013},
+  urldate = {2021-04-23},
   langid = {english}
 }
 
@@ -205,6 +218,7 @@
   pages = {W11422},
   issn = {1944-7973},
   doi = {10.1029/2005WR004241},
+  urldate = {2017-05-17},
   langid = {english}
 }
 
@@ -219,6 +233,7 @@
[…]
