### 2023-05-10 #5427 @elbeejay
Thanks @mgcooper for providing those answers above and that short-list of potential reviewers.

My one suggestion (at this time) for the paper is the removal of the "Software Availability" section, as the left side of the PDF, as well as the JOSS webpage will have links to the software repository as well as the archived version of the code.

I'll start reaching out to potential reviewers here next
