### 2023-05-10 #5427 @mgcooper
Hi @Lvulis (submitting author here, so this is not an official reply) much of the package will run in Octave but overall I don't think the paper can be reviewed using Octave at this time. I could probably get it working by early next week, but I was hoping to defer this until after May 31. The main issue is with string and datetime compatibility, but the live scripts need to be converted to regular scripts as well (unless live scripts are supported in Octave or can be converted to Octave equivalent notebooks). I also need to make a quick edit to the non-linear curve fitting routine to ensure compatibility with Octave's optim package.

@elbeejay, if Octave compatibility ends up being necessary to get the paper into the review process, please let me know and I will do my best to expedite that. Many thanks,

Matt
