### 2023-05-11 #5427 @elbeejay
Thanks @Lvulis and @DavidLitwin for the replies. I actually don't have a matlab license either so I totally get it. @mgcooper there is no requirement for your software to be Octave compatible, but we'd want the paper to accurately describe what languages are supported - so it is up to you if you'd like to support Octave
