### 2023-05-12 #5427 @elbeejay
@deboraydo would you be interested in reviewing this submission to JOSS titled: "baseflow: a MATLAB and GNU Octave package for baseflow recession analysis"? Lawrence recommended you as a good fit for this submission. I've included a summary of the JOSS review process below.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission - I am aware some of you may have contributed to the ObsPy package to which @trichter is also a contributor, so if you are willing to review and have contributed there I would ask that you simply state that fact here, so we are as clear and transparent as possible. It is such a large community project that I do not view being co-contributors to it as a true conflict of interest unless you've worked closely together to develop a specific feature, for example.

Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know!

Thanks,
Jay
