### 2023-05-12 #5427 @elbeejay
Thanks for the suggestion @amabdallah! @aymnassar would you be interested in doing this review for JOSS? Please reach out if interested or if you have any questions about the process.
