### 2023-05-12 #5427 @deboraydo
Dear Jay,

Thank you for the invitation. I am interested in reviewing this submission.

Thank you,
Debora

On Fri, May 12, 2023 at 9:56 AM J. Hariharan ***@***.***>
wrote:

> Thanks for the suggestion @amabdallah <https://github.com/amabdallah>!
> @aymnassar <https://github.com/aymnassar> would you be interested in
> doing this review for JOSS? Please reach out if interested or if you have
> any questions about the process.
>
