### 2023-05-15 d92c57d
chore: rename LICENSE to COPYING and add DESCRIPTION for octave pkg format
