### 2023-05-17 #5427 @mgcooper
@elbeejay I fixed the main Octave incompatibilities but I have not pushed them to joss branch. Please let me know the  protocol regarding changes, I reviewed the joss docs and was not able to find it. I understand the issues page is used to raise / address issues, but is there an expectation that the review branch will be "frozen" at some point while the referees complete their reviews? I also wanted to make sure that the referees review the joss branch, not the main branch (or dev). Many thanks.
