### 2023-05-18 #5427 @kellykochanski
Hi all,

I'm afraid I have some heavy work weeks coming up, and can't commit to
providing timely feedback during the review process.

As with Chad, if you haven't found someone again by mid-June, please feel
free to ask again :)

On Wed, May 17, 2023 at 10:21 AM Matt Cooper ***@***.***>
wrote:

> @elbeejay <https://github.com/elbeejay> I fixed the main Octave
> incompatibilities but I have not pushed them to joss branch. Please let me
> know the protocol regarding changes, I reviewed the joss docs and was not
> able to find it. I understand the issues page is used to raise / address
> issues, but is there an expectation that the review branch will be "frozen"
> at some point while the referees complete their reviews? I also wanted to
> make sure that the referees review the joss branch, not the main branch (or
> dev). Many thanks.
>
