### 2023-05-19 #5427 @elbeejay
Thanks @chadagreene, @fgnievinski, and @kellykochanski for letting us know your situation and future availability. We definitely appreciate it.

@fgnievinski if you are comfortable reviewing in a few weeks, that would work with the typical JOSS review period which is as long as 6 weeks.

@mgcooper - thank you for implementing the Octave compatibility fixes. JOSS reviews are very fluid, so please feel free to push to your project as needed. Regarding the review taking place in the `joss` branch, that is fine and we can communicate that to reviewers.
