### 2023-05-23 #5492 @elbeejay
Hi @deboraydo, @aymnassar, and @alessandroamaranto,

This is the official "review" issue. @mgcooper has asked that you review the content in the `joss` branch of the `baseflow` repository, so please make your comments accordingly. Instructions for creating your reviewer checklist and conducting the review should be in the top comment of this issue, but please do not hesitate to reach out to me with any questions or to ask for any clarification.

Ideally we'd like to ask that you complete your reviews within 6 weeks, and I will set up reminder so the bot prods us all in 3 weeks.

Thanks,
Jay
