### 2023-06-17 cfb7d30
feat(octave): octave compatibility replace ax objects with set,get, contains with ismember, magicparser with inputparser, string == with strcmp
new functions nanmovmean, emptyaxes, islocalmax, latex2tex, fitNLS_octave

add persistent variable inoctave, replace latex with if inoctave tex

update example demos for octave compatibility
