+++ +bfra/+deps/peakfinder.m
@@ -1,0 +1,371 @@
+function [peakInds,peakMags] = peakfinder(x0, sel, thresh, extrema, includeEndpoints, interpolate)
+%PEAKFINDER Noise tolerant fast peak finding algorithm
+%   INPUTS:
+%       x0 - A real vector from the maxima will be found (required)
+%       sel - The amount above surrounding data for a peak to be,
+%           identified (default = (max(x0)-min(x0))/4). Larger values mean
+%           the algorithm is more selective in finding peaks.
+%       thresh - A threshold value which peaks must be larger than to be
+%           maxima or smaller than to be minima.
+%       extrema - 1 if maxima are desired, -1 if minima are desired
+%           (default = maxima, 1)
+%       includeEndpoints - If true the endpoints will be included as
+%           possible extrema otherwise they will not be included
+%           (default = true)
+%       interpolate - If true quadratic interpolation will be performed
+%           around each extrema to estimate the magnitude and the
+%           position of the peak in terms of fractional indicies. Note that
+%           unlike the rest of this function interpolation assumes the
+%           input is equally spaced. To recover the x_values of the input
+%           rather than the fractional indicies you can do:
+%           peakX = x0 + (peakLoc - 1) * dx
+%           where x0 is the first x value and dx is the spacing of the
+%           vector. Output peakMag to recover interpolated magnitudes.
+%           See example 2 for more information.
+%           (default = false)
+%
+%   OUTPUTS: mgc switched order back to loc,mag (mag,loc in +pfa)
+%       peakMag - The magnitude of the identified peaks
+%       peakLoc - The indicies of the identified peaks in x0
+%
+%   [peakLoc] = peakfinder(x0) returns the indicies of local maxima that
+%       are at least 1/4 the range of the data above surrounding data.
+%
+%   [peakLoc] = peakfinder(x0,sel) returns the indicies of local maxima
+%       that are at least sel above surrounding data.
+%
+%   [peakLoc] = peakfinder(x0,sel,thresh) returns the indicies of local
+%       maxima that are at least sel above surrounding data and larger
+%       (smaller) than thresh if you are finding maxima (minima).
+%
+%   [peakLoc] = peakfinder(x0,sel,thresh,extrema) returns the maxima of the
+%       data if extrema > 0 and the minima of the data if extrema < 0
+%
+%   [peakLoc] = peakfinder(x0,sel,thresh,extrema, includeEndpoints)
+%       returns the endpoints as possible extrema if includeEndpoints is
+%       considered true in a boolean sense
+%
+%   [peakLoc, peakMag] = peakfinder(x0,sel,thresh,extrema,interpolate)
+%       returns the results of results of quadratic interpolate around each
+%       extrema if interpolate is considered to be true in a boolean sense
+%
+%   [peakLoc, peakMag] = peakfinder(x0,...) returns the indicies of the
+%       local maxima as well as the magnitudes of those maxima
+%
+%   If called with no output the identified maxima will be plotted along
+%       with the input data.
+%
+%   Note: If repeated values are found the first is identified as the peak
+%
+% Example 1:
+% t = 0:.0001:10;
[…]
