+++ +bfra/+validation/isaxis.m
@@ -1,3 +1,7 @@
 function tf = isaxis(ax)
 %ISAXIS logical check if axis object
-tf = isa(ax,'matlab.graphics.axis.Axes');
+if bfra.util.isoctave
+   tf = isaxes(ax);
+else
+   tf = isa(ax,'matlab.graphics.axis.Axes');
+end
