+++ demos/bfra_demo_2.m
@@ -1,0 +1,143 @@
+%% Estimating recession parameters with the Pareto distribution method
+%% Introduction
+% This example demonstrates how to detect recession events, and then estimate
+% recession parameters _a_ and _b_ using the hybrid Pareto distribution-"point
+% cloud" method (Cooper et al. 2023). Unlike the traditional point-cloud method,
+% the key advantage of this method is that the Pareto distribution identifies
+% "late-time" drainage behavior and provides an estimate of _b_ for the late-time
+% regime.
+%
+% We use daily streamflow data measured in the Kuparuk River Basin on the North
+% Slope of Alaska provided by the United States Geological Survey, recorded at
+% streamflow gage 1596000. The data are included in the |data/| folder as a sample
+% dataset with this toolbox, or can be downloaded from <https://waterdata.usgs.gov/monitoring-location/15896000.
+% https://waterdata.usgs.gov/monitoring-location/15896000.>
+%%
+%% Preparing data for recession analysis
+%
+%
+% Set the sitename.
+
+sitename = bfra.basinname('KUPARUK R NR DEADHORSE AK');
+%%
+% Load streamflow data for the test basin into the workspace. In the sample
+% dataset, the variable |T| is _time_, |Q| is _discharge_, and |R| is _rainfall_.
+
+[T,Q,R] = bfra.test.testdata('example');
+%%
+% Set the basin area and geomorphological parameters.
+
+Dd = 0.8;            % drainage density
+A = 8.6545e9;        % basin area [m2]
+L = A*Dd/1000;       % active stream length
+D = 0.5;             % saturated aquifer thickness
+%%
+%% Set options for detecting and fitting recession events
+%
+%
+% The toolbox supports three main tasks: recession event detection, recession
+% event curve-fitting, and parameter estimation via distribution fitting. The
+% |setopts| function provides an interface to set options for the algorithms that
+% implement these tasks.
+%
+% Set default algorithm options for |getevents| and |fitevents|.
+
+opts.getevents = bfra.setopts('getevents');
+opts.fitevents = bfra.setopts('fitevents');
+%%
+%
+%
+% In addition to |getevents and| |fitevents|, in this example we set options
+% for |globalfit|. The term "globalfit" differentiates two fitting procedures:
+% one that fits individual recession events to estimate parameters specific to
+% each  particular event (which is the purpose of |fitevents|), and another that
+% fits a collection of events to derive a "global" parameter value applicable
+% to the entire sample population (the purpose of |globalfit|).
+%
+% Set options for |globalfit.|
+
+opts.globalfit = bfra.setopts('globalfit','drainagearea',A,'streamlength',L, ...
+   'drainagedens',Dd,'aquiferdepth',D,'isflat',true,'plotfits',true);
+%%
+%% Recession analysis workflow
+%
+%
+% The following steps demonstrate a typical workflow in baseflow recession analysis.
+% The steps consist of getting a list of recession events, fitting the recession
+% events, and then estimating the recession parameters _a_ and _b_.
+% Step 1. Get events
+
+EventData = bfra.getevents(T,Q,R,opts.getevents);
+%
+% Step 2. Fit events
+
+[EventFits,FitsTable] = bfra.fitevents(EventData,opts.fitevents);
+%
+% Step 3. Fit recession parameters _a_ and _b_
+%
+%
+% The |globalfit| function fits a Pareto distribution to the sample population
+% of event-scale parameters stored in |FitsTable|. The parameters of the Pareto
+% distribution $\tau_0$ and $\alpha$ can be related to the traditional recession
+% analysis parameters $a$ and $b$. For details, see Cooper et al. 2023. The value
[…]
