+++ README.md
@@ -8,10 +8,11 @@ A matlab toolbox for characterizing aquifer properties from streamflow measureme
 
 ## Dependencies
 
-Developed on MATLAB Version: 9.9.0 (R2020b)  
-Mathworks Statistics and Machine Learning Toolbox&trade;  
-Mathworks Curve Fitting Toolbox&trade;  
-Various utility functions included in this toolbox
+Developed on MATLAB Version: 9.9.0 (R2020b).  
+Mathworks Statistics and Machine Learning Toolbox&trade;.  
+Mathworks Curve Fitting Toolbox&trade;.  
+Compatible with Octave, tested on version 8.2.0.  
+Requires Statistics, Optimization, Struct, and Tablicious packages.  
 
 ## Install
 
@@ -38,7 +39,7 @@ Toolbox documentation and examples are also available in the Matlab help browser
 - Type `doc bfra` or try `doc +bfra` to see the package contents in the help browser.
 - If the documentation does not open in the help browser, try `doc` without any arguments, then scroll down to "Supplemental Software" and click on "Baseflow Recession Analysis Toolbox". You can also try `docsearch bfra`.
 
-In addition to the toolbox documentation and examples provided in the Matlab help browser, the notebook in `demos/bfra_kuparuk.mlx` replicates the analysis in the paper [Detecting Permafrost Active Layer Thickness Change From Nonlinear Baseflow Recession](https://doi.org/10.1029/2022WR033154). There are also several other notebooks in the `demos` folder that demonstrate how to use the toolbox functions, each of which are also available as html files in the `docs` folder, and are viewable in the Matlab help browser.
+In addition to the toolbox documentation and examples provided in the Matlab help browser, there are several notebooks in `demos/`. The notebook `bfra_kuparuk.mlx` replicates the analysis in the paper [Detecting Permafrost Active Layer Thickness Change From Nonlinear Baseflow Recession](https://doi.org/10.1029/2022WR033154). Several other notebooks in the `demos` folder demonstrate how to use the toolbox functions, each of which are also available as html files in the `docs` folder, and are viewable in the Matlab help browser. All demos are available as live `.mlx` files compatible with Matlab, and as `.m` files compatible with Octave.
 
 ## Contribute
 
@@ -78,7 +79,7 @@ For extended use of the toolbox, the convenience function `Setup.m` includes opt
   - `Setup('rmpath')` removes the toolbox from the search path for the current session.
 - To display the current toolbox preferences try `getpref('baseflow')`.
 
-Running `Setup('install')` should only be necessary once (or not at all, if you choose to manage your search path however you normally manage third-party matlab/octave software). However, Octave users may find Setup particularly convenient because Setup will load the required packages (but see [Octave](#octave) for compatibility note). Although all dependencies are nominally included in this toolbox, if users encounter any missing dependencies, please open an [issue](https://github.com/mgcooper/baseflow/issues).
+Running `Setup('install')` should only be necessary once (or not at all, if you choose to manage your search path however you normally manage third-party matlab/octave software). However, Octave users may find Setup particularly convenient because Setup will load the required packages. Although all dependencies are nominally included in this toolbox, if users encounter any missing dependencies, please open an [issue](https://github.com/mgcooper/baseflow/issues).
 
 <!-- Disabled this after moving all dependencies to package namespace folders and running built in matlab dependency report -->
  <!-- If for some reason a dependencies is found that is not on the search path, a message is printed to the screen. To see the list of missing dependencies, check the `msg` output. At any time, a dependencies check can be run using: -->
@@ -102,14 +103,14 @@ If desired, package functions can be imported into a workspace using `import bfr
 
 ## Octave
 
[…]
