### 2023-07-03 #5492 @alessandroamaranto
The package is easy to install, use and understand. It is well-documented and comes with a set of workflows to ease the users through the functionalities. Much appreciated, thanks.
