### 2023-07-21 3a03fae
fix(tests): fix size mismatch in test_eventfinder
remove non-functional geteventsTest

add test_islocalmax

add verbose and debugging to runtests
