### 2023-07-21 9afd83d
refactor(parsing): make parsers persistent and use bfra.validation.is* functions
