+++ +bfra/fitevents.m
@@ -34,45 +34,18 @@ function [Fits,Results] = fitevents(Events,varargin)
 % if called with no input, open this file
 if nargin == 0; open(mfilename('fullpath')); return; end
 
-%-------------------------------------------------------------------------------
-p = inputParser;
-p.FunctionName = 'fitevents';
-p.StructExpand = true;
-% p.PartialMatching = true;
+% PARSE INPUTS
 
-addRequired(p, 'Events',               @(x) isstruct(x)                 );
-addParameter(p,'derivmethod', 'ETS',   @(x) ischar(x)                   );
-addParameter(p,'fitmethod',   'nls',   @(x) ischar(x)                   );
-addParameter(p,'fitorder',    'free',  @(x) ischar(x) | isnumeric(x)    );
-addParameter(p,'pickfits',    false,   @(x) islogical(x) & isscalar(x)  );
-addParameter(p,'pickmethod',  'none',  @(x) ischar(x)                   );
-addParameter(p,'plotfits',    false,   @(x) islogical(x) & isscalar(x)  );
-addParameter(p,'saveplots',   false,   @(x) islogical(x) & isscalar(x)  );
-addParameter(p,'etsparam',    0.2,     @(x) isnumeric(x) & isscalar(x)  );
-addParameter(p,'vtsparam',    1.0,     @(x) isnumeric(x) & isscalar(x)  );
+[derivmethod, fitmethod, fitorder, pickfits, pickmethod, plotfits, saveplots, ...
+   etsparam, vtsparam, fitopts] = parseinputs(Events, mfilename, varargin{:});
 
+% MAIN FUNCTION
 
-parse(p,Events,varargin{:});
-
-derivmethod = p.Results.derivmethod;
-fitmethod   = p.Results.fitmethod;
-fitorder    = p.Results.fitorder;
-pickfits    = p.Results.pickfits;
-pickmethod  = p.Results.pickmethod;
-plotfits    = p.Results.plotfits;
-saveplots   = p.Results.saveplots;
-etsparam    = p.Results.etsparam;
-vtsparam    = p.Results.vtsparam;
-fitopts     = struct(); % removed fitopts from parser, it may mess up autounpacking
-%-------------------------------------------------------------------------------
-
-% pull out the events
-T           =  Events.eventTime;       % time [days]
-Q           =  Events.eventFlow;       % daily discharge [m3 d-1] 
-R           =  Events.eventRain;       % daily rainfall [mm d-1]
-eventTags   =  Events.eventTags;
-numEvents   =  max(eventTags);
-ax          =  'none';
+eventTime = Events.eventTime;               % time [days]
+eventFlow = Events.eventFlow;               % daily discharge [m3 d-1] 
+eventRain = Events.eventRain;               % daily rainfall [mm d-1]
+eventTags = Events.eventTags;
+numEvents = max(eventTags);
 
 % try
 %    T = datetime(T,'ConvertFrom','datenum');
@@ -80,16 +53,15 @@ ax          =  'none';
 % end
 
 % initialize output structure and output arrays
-Fits.eventTime =  Events.eventTime;    % event-times
-Fits.eventFlow =  Events.eventFlow;    % detected event-Q
-% Fits.t         =  NaT(size(Q));        % fitted t
-Fits.t         =  nan(size(Q));        % fitted t
-Fits.q         =  nan(size(Q));        % fitted Q
-Fits.r         =  nan(size(Q));        % rain
-Fits.dt        =  nan(size(Q));        % fitted dt
[…]
