+++ +bfra/getdqdt.m
@@ -5,7 +5,7 @@ function [q,dqdt,dt,tq,rq,varargout] = getdqdt(T,Q,R,derivmethod,varargin)
 %
 %     [q,dqdt,dt,tq,rq] = bfra.getdqdt(T,Q,R,derivmethod)
 %     [q,dqdt,dt,tq,rq] = bfra.getdqdt(_,'fitwindow',fitwindow)
-%     [q,dqdt,dt,tq,rq] = bfra.getdqdt(_,'fitwindow',fitmethod)
+%     [q,dqdt,dt,tq,rq] = bfra.getdqdt(_,'fitmethod',fitmethod)
 %     [q,dqdt,dt,tq,rq] = bfra.getdqdt(_,'pickmethod',pickmethod)
 %     [q,dqdt,dt,tq,rq] = bfra.getdqdt(_,'ax',axis_object)
 %
@@ -31,48 +31,22 @@ function [q,dqdt,dt,tq,rq,varargout] = getdqdt(T,Q,R,derivmethod,varargin)
 %     fitab    =  logical, scalar, indicates whether to fit a/b in -dQ/dt=aQb
 %     plotfit  =  logical, scalar, indicates whether to plot the fit
 %
-% See also getdqdt
+% See also fitdqdt
 %
 % Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
 
-% if called with no input, open this file
-if nargin == 0; open(mfilename('fullpath')); return; end
-
-%
 % Tip: this accepts pre-selected events, not raw timeseries. Use
 % bfra.findevents to pick Events, then bfra.getdqdt to fit the events.
 % This is a wrapper for multi-year, final analysis.
 
-%-------------------------------------------------------------------------------
-% input handling
-p                 = inputParser;
-p.FunctionName    = 'getdqdt';
-p.CaseSensitive   = true;
+% if called with no input, open this file
+if nargin == 0; open(mfilename('fullpath')); return; end
 
-addRequired(p, 'T',                    @(x) isnumeric(x) | isdatetime(x));
-addRequired(p, 'Q',                    @(x) isnumeric(x)                );
-addRequired(p, 'R',                    @(x) isnumeric(x)                );
-addRequired(p, 'derivmethod',          @(x) ischar(x)                   );
-addParameter(p,'vtsparam',    1,       @(x) isnumeric(x) & isscalar(x)  );
-addParameter(p,'etsparam',    0.2,     @(x) isnumeric(x) & isscalar(x)  );
-addParameter(p,'ctsmethod',   'B1',    @(x) ischar(x)                   );
-addParameter(p,'pickmethod',  'none',  @(x) ischar(x)                   );
-addParameter(p,'fitmethod',   'nls',   @(x) ischar(x)                   );
-addParameter(p,'plotfits',    false,   @(x) islogical(x) & isscalar(x)  );
-addParameter(p,'eventID',     'none',  @(x) ischar(x)                   );
+% PARSE INPUTS
+[vtsparam, etsparam, ctsmethod, pickmethod, fitmethod, plotfits, eventID] = ...
+   parseinputs(T,Q,R,derivmethod,mfilename,varargin{:});
 
-parse(p,T,Q,R,derivmethod,varargin{:});
-
-vtsparam    = p.Results.vtsparam;
-etsparam    = p.Results.etsparam;
-ctsmethod   = p.Results.ctsmethod;
-pickmethod  = p.Results.pickmethod;
-fitmethod   = p.Results.fitmethod;
-plotfits    = p.Results.plotfits;
-eventID     = p.Results.eventID;
-
-%-------------------------------------------------------------------------------
-
+% MAIN FUNCTION
 switch derivmethod
 
[…]
