+++ +bfra/+validation/isdatelike.m
@@ -1,2 +1,2 @@
 function tf = isdatelike(x)
-tf = isdatetime(x) | isnumeric(x) & isvector(x);
+tf = isdatetime(x) || (isnumeric(x) && isvector(x));

[…]
