+++ paper.bib
@@ -251,15 +251,3 @@
   doi = {10.1002/wrcr.20407},
   urldate = {2021-03-10}
 }
-@article{cooper:2023,
-  url = {https://onlinelibrary.wiley.com/doi/pdf/10.1029/2022WR033154},
-  Author = {{Cooper}, M.~G. and {Zhou}, T. and {Bennett}, K.~E. and {Bolton}, W.~R. and {Coon}, E.~T. and {Fleming}, S.~W. and {Rowland}, J.~C. and {Schwenk}, J.},
-  Journal = {Water Resources Research},
-  Title = {{Detecting Permafrost Active Layer Thickness Change From Nonlinear Baseflow Recession}},
-  Year = 2023,
-  volume = 59,
-  number = 1,
-  pages = {e2022WR033154},
-  issn = {1944-7973},
-  doi = {10.1029/2022WR033154}
-}
