### 2023-08-03 5b67285
refactor(parser): move all parsers to local function
