### 2023-08-16 0a10b28
feat(test): add eventfinder rmconvex test
use clearer notation and style in eventsplitter
