+++ +bfra/+deps/peakfinder.m
@@ -78,23 +78,28 @@ function [peakInds,peakMags] = peakfinder(x0, sel, thresh, extrema, includeEndpo
 % set(p(2:end), 'linewidth', 2); % Show the markers more clearly
 % legend('Actual Data', 'Input Data', 'Estimated Peaks');
 % Copyright Nathanael C. Yoder 2015 (nyoder@gmail.com)
-% 
+%
 % mgc: 'sel' = 'Threshold', 'thresh' = 'MinPeakHeight'
-
+% peakfinder(X, sel, thresh, extrema, includeEndpoints, interpolate)
+% peakfinder(X, 0, 0, 1, false) is equivalent to:
+% findpeaks(X, 'Threshold', 0, 'MinPeakHeight', 0)
+% arguments not included in the call to findpeaks: 1 for maxima (-1 would be
+% minima), false for includeEndpoints.
+%
 % ------------------------------------------------------------------------
 % Copyright (c) 2016, Nathanael C. Yoder
 % All rights reserved.
-% 
+%
 % Redistribution and use in source and binary forms, with or without
 % modification, are permitted provided that the following conditions are
 % met:
-% 
+%
 %     * Redistributions of source code must retain the above copyright
 %       notice, this list of conditions and the following disclaimer.
 %     * Redistributions in binary form must reproduce the above copyright
 %       notice, this list of conditions and the following disclaimer in
 %       the documentation and/or other materials provided with the distribution
-% 
+%
 % THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 % AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 % IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -117,55 +122,55 @@ s = size(x0);
 flipData =  s(1) < s(2);
 len0 = numel(x0);
 if len0 ~= s(1) && len0 ~= s(2)
-    error('PEAKFINDER:Input','The input data must be a vector')
+   error('PEAKFINDER:Input','The input data must be a vector')
 elseif isempty(x0)
-    varargout = {[],[]};
-    return;
+   varargout = {[],[]};
+   return;
 end
 if ~isreal(x0)
-    warning('PEAKFINDER:NotReal','Absolute value of data will be used')
-    x0 = abs(x0);
+   warning('PEAKFINDER:NotReal','Absolute value of data will be used')
+   x0 = abs(x0);
 end
 
 if nargin < 2 || isempty(sel)
-    sel = (max(x0)-min(x0))/4;
+   sel = (max(x0)-min(x0))/4;
 elseif ~isnumeric(sel) || ~isreal(sel)
-    sel = (max(x0)-min(x0))/4;
-    warning('PEAKFINDER:InvalidSel',...
-        'The selectivity must be a real scalar.  A selectivity of %.4g will be used',sel)
+   sel = (max(x0)-min(x0))/4;
+   warning('PEAKFINDER:InvalidSel',...
+      'The selectivity must be a real scalar.  A selectivity of %.4g will be used',sel)
 elseif numel(sel) > 1
-    warning('PEAKFINDER:InvalidSel',...
-        'The selectivity must be a scalar.  The first selectivity value in the vector will be used.')
-    sel = sel(1);
+   warning('PEAKFINDER:InvalidSel',...
+      'The selectivity must be a scalar.  The first selectivity value in the vector will be used.')
+   sel = sel(1);
 end
 
 if nargin < 3 || isempty(thresh)
-    thresh = [];
+   thresh = [];
[…]
