+++ +bfra/+test/ParameterizedTestBfra.m
@@ -12,6 +12,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       PhiValue = {0.001, 0.01, 0.1};
       BasinName = {'ALL_BASINS','KUPARUK R NR DEADHORSE AK'};
       MinEventDuration = {3,6,9};
+      RmConvex = {false, true};
    end
 
    methods (Test)
@@ -35,7 +36,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       % Verify that the actual result matches the expected result
       testCase.verifyEqual(strActual,expectedStr)
       end
-   
+
       %-------------------------------------------
       %-------------------------------------------
       function test_getdqdt(testCase,DerivMethod)
@@ -53,7 +54,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
 
       % Get the actual result
       [~,dqdtActual] = bfra.getdqdt(t,q,[],DerivMethod);
-         
+
       % Verify that the actual result matches the expected result
       testCase.verifyEqual(dqdtActual,dqdtExpected)
 
@@ -68,7 +69,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       %-------------------------------------------
       %-------------------------------------------
       function test_fitab(testCase,RecessionExponent,FitMethod)
-      
+
       % define values to generate the test data
       a = 1e-2;
       q0 = 100;
@@ -81,8 +82,8 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       % [q,dqdt] = bfra.util.generateTestData(a,b,q0,t);
       % figure; loglog(q,-dqdt,'o')
       % also useful to see this:
-      % bfra.Qnonlin(a,b,q0,t,true) 
-      
+      % bfra.Qnonlin(a,b,q0,t,true)
+
       % f = bfra.fitab(q,dqdt,'mean'); ab = f.ab
       % f = bfra.fitab(q,dqdt,'ols'); ab = f.ab
       % f = bfra.fitab(q,dqdt,'nls'); ab = f.ab
@@ -98,7 +99,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
          otherwise
             Fit = bfra.fitab(q,dqdt,FitMethod);
       end
-      
+
       % Get the expected result
       abExpected = [a;b];
 
@@ -108,7 +109,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       % for testing:
       % f = bfra.fitab(q,dqdt,FitMethod); abActual = f.ab
       % isequal(abActual,abExpected)
-         
+
       % Verify that the actual result matches the expected result
       testCase.verifyEqual(abActual,abExpected,'RelTol',[0.01; 0.01])
 
@@ -123,7 +124,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
             b = 1.5;
             nExpected = 0;
             nActual = bfra.conversions(b,'b','n','isflat',true);
-            
+
             % Verify that the actual result matches the expected result
             testCase.verifyEqual(nActual,nExpected);
 
@@ -181,7 +182,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
             SminExpected = 1/a/(2-b).*(qmin.^(2-b));
             SmaxExpected = 1/a/(2-b).*(qmax.^(2-b));
       end
-      
+
       % get the actual result
       [SminActual,SmaxActual] = bfra.aquiferstorage(a,b,qmin,qmax);
 
@@ -205,7 +206,7 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       % compute the expected result
[…]

[…]
