### 2023-08-16 0157e44
fix(loaddata): fix parsers, replace setpath with fullfile to BASEFLOW_DATA_PATH, replace tocol with tocolumn
add dS/dt to getstring
