### 2023-08-16 fe5e23e
feat(mapbasins): add custom colormap and climits
