+++ docs/makedocs.m
@@ -7,24 +7,26 @@ function makedocs(varargin)
 % three options: publish example docs, function docs, or build docsearch db
 opts = bfra.util.optionParser({'examples','functions','docsearch'},varargin(:));
 
+% set paths 
 thispath = fileparts(mfilename('fullpath'));
+basepath = fileparts(thispath);
 
 testrun = false;
 if testrun == true
-   docspath = fullfile(thispath,'testbed');
+   docspath = fullfile(basepath, 'testbed', 'docs');
 else
-   docspath = fullfile(thispath,'docs');
+   docspath = thispath;
 end
 
 % nest m2html stuff in html/ to easily distinguish it from matlab publish stuff
-htmlpath = fullfile(docspath,'html');
-m2htmlpath = fullfile(htmlpath,'m2html');
+htmlpath = fullfile(docspath, 'html');
+m2htmlpath = fullfile(htmlpath, 'm2html');
 
 % matlab publish opts
 mpubopts = struct( ...
-   'format','html', ...
-   'outputDir',htmlpath, ...
-   'useNewFigure',false ...
+   'format', 'html', ...
+   'outputDir', htmlpath, ...
+   'useNewFigure', false ...
    );
 
 %% build a doc search database
@@ -38,15 +40,15 @@ end
 if opts.examples == true
    
    % publish the standard docs using matlab's publish
-   publish(fullfile(docspath,'bfra_welcome.m'),mpubopts);
-   publish(fullfile(docspath,'bfra_gettingStarted.m'),mpubopts);
-   % publish(fullfile(docspath,'bfra_example_1.m'),mpubopts);
-   publish(fullfile(docspath,'bfra_contents.m'),mpubopts);
-   publish(fullfile(docspath,'bfra_theory_contents.m'),mpubopts);
-   publish(fullfile(docspath,'bfra_examples_contents.m'),mpubopts);
+   publish(fullfile(docspath, 'bfra_welcome.m'), mpubopts);
+   publish(fullfile(docspath, 'bfra_gettingStarted.m'), mpubopts);
+   publish(fullfile(docspath, 'bfra_contents.m'), mpubopts);
+   publish(fullfile(docspath, 'bfra_theory_contents.m'), mpubopts);
+   publish(fullfile(docspath, 'bfra_examples_contents.m'), mpubopts);
 
    % for github actions the landing page must be saved as index.html
-   copyfile(fullfile(htmlpath,'bfra_gettingStarted.html'),fullfile(docspath,'index.html'));
+   copyfile(fullfile(htmlpath, 'bfra_gettingStarted.html'), ...
+      fullfile(docspath, 'index.html'));
 
    % not sure if other files can also be supported, if so:
    % copyfile('html/bfra_demo.html','bfra_demo.html');
@@ -57,10 +59,17 @@ end
 if opts.functions == true
    
    % activate m2html
-   activate m2html_rochefort
+   try
+      activate m2html_rochefort
+   catch e
+      rethrow(e)
+   end
+   
+   % run m2html from the project base directory (one dir above this one)
+   job = withcd(basepath); %#ok<NASGU> 
 
    % note that m2html uses relative paths. generate a list of dirs to ignore.
-   alldirs = dir([thispath filesep]); alldirs = alldirs([alldirs.isdir]);
+   alldirs = dir([basepath filesep]); alldirs = alldirs([alldirs.isdir]);
    alldirs(strncmp({alldirs.name}, '.', 1) & strlength({alldirs.name})<3) = [];
    ignored = {alldirs(~ismember({alldirs.name},'+bfra')).name};
 
@@ -72,18 +81,18 @@ if opts.functions == true
    end
 
    m2html( ...
-      'mfiles',mfilepath, ...             % source dir where the files live
[…]
