### 2023-08-21 #5492 @tianydong
Overall the package is easy to install, the codes are well-commented, and demo 1 runs fine.
But I am getting some errors on the other demos, for example:
![image](https://github.com/openjournals/joss-reviews/assets/72535053/61c0c83b-f26b-45e1-a2d8-182d704ffeff)

Maybe a path thing? Also, I am using MATLAB 2022b, but I wonder if that's a problem.
Also, there's a conflict between the .m and .mlx files for the linear theory demos.
![image](https://github.com/openjournals/joss-reviews/assets/72535053/d46d2a4b-97f0-4ca8-a5ef-45f242b7557b)
does this mean the user suppose to stick with LiveScript?
Thanks again for having me as part of the review.
I will provide some comments on the paper shortly
