### 2023-08-21 #5492 @tianydong
I made some minor comments, but overall the paper is well-written, and the figure is nicely made. Good luck!
[paper_TDedits.pdf](https://github.com/openjournals/joss-reviews/files/12399502/paper_TDedits.pdf)
