### 2023-08-25 #5492 @mgcooper
@elbeejay @tianydong @alessandroamaranto

Thank you kindly for the review, @tianydong and @alessandroamaranto! I will formally respond to the edits and make the needed code revisions to resolve the errors, but wanted to quickly explain them. The first one occurs because the input parser allows the argument T to be either double (for datenum compatibility) or datetime. But then I have a second check (the one throwing the error, on line 180) which is requiring T to be double. So I will just add 'datetime' to that validation function. This is a newer change I pushed recently, so it did not affect @alessandroamaranto version.

The second error is because Matlab does not allow m-files (.m) to have the same name as live scripts (.mlx). To work around this, I used "demo" for the .m files and "example" for the .mlx. For example, "bfra_demo_1.m" is identical to "bfra_example_1.mlx" in terms of the content, but the former is a traditional .m file and the latter is in the newer live script format. I just completely overlooked this name shadowing for the two "theory" examples. So I'll rename them for the revisions. You could also change the name to anything other than the current one  and it should work (as long as the new name doesn't shadow some other file) if you want to try running it.

I'll get these fixes pushed asap. Thanks again for the reviews and for the edits everyone,
Matt
