### 2023-08-25 fd7af38
refactor: move flattenevents to private
