+++ +bfra/getevents.m
@@ -115,7 +115,7 @@ end
       Events.eventTime, ...
       Events.eventFlow, ...
       Events.eventRain, ...
-      Events.eventTags] = bfra.flattenevents(t,q,r,Info);
+      Events.eventTags] = flattenevents(t,q,r,Info);
 % end
 
 
