### 2023-08-25 f78c2ee
feat(test-setEventEmpty): add test_MinEventDuration to implicitly test setEventEmpty
move setEventEmpty to private
