+++ +bfra/+test/ParameterizedTestBfra.m
@@ -21,48 +21,48 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       %-------------------------------------------
       function test_getstring(testCase,VarStr)
 
-      switch VarStr
-         case 'Q'
-            expectedStr = '$Q$';
-         case 'dQdt'
-            expectedStr = '$-\mathrm{d}Q/\mathrm{d}t$';
-         case 'aQb'
-            expectedStr = '$-\mathrm{d}Q/\mathrm{d}t = aQ^b$';
-      end
+         switch VarStr
+            case 'Q'
+               expectedStr = '$Q$';
+            case 'dQdt'
+               expectedStr = '$-\mathrm{d}Q/\mathrm{d}t$';
+            case 'aQb'
+               expectedStr = '$-\mathrm{d}Q/\mathrm{d}t = aQ^b$';
+         end
 
-      % Get actual result
-      strActual = bfra.getstring(VarStr);
+         % Get actual result
+         strActual = bfra.getstring(VarStr);
 
-      % Verify that the actual result matches the expected result
-      testCase.verifyEqual(strActual,expectedStr)
+         % Verify that the actual result matches the expected result
+         testCase.verifyEqual(strActual,expectedStr)
       end
 
       %-------------------------------------------
       %-------------------------------------------
       function test_getdqdt(testCase,DerivMethod)
 
-      % define the test data
-      a = 1e-2;
-      b = 1.5;
-      q0 = 1;
-      [q,~,t] = bfra.util.generateTestData(a,b,q0);
+         % define the test data
+         a = 1e-2;
+         b = 1.5;
+         q0 = 1;
+         [q,~,t] = bfra.util.generateTestData(a,b,q0);
 
-      % Calculate expected result for CTS method
-      dq = q-[nan; q(1:end-1)];
-      dt = (t(2)-t(1));
-      dqdtExpected = dq./dt;
+         % Calculate expected result for CTS method
+         dq = q-[nan; q(1:end-1)];
+         dt = (t(2)-t(1));
+         dqdtExpected = dq./dt;
 
-      % Get the actual result
-      [~,dqdtActual] = bfra.getdqdt(t,q,[],DerivMethod);
+         % Get the actual result
+         [~,dqdtActual] = bfra.getdqdt(t,q,[],DerivMethod);
 
-      % Verify that the actual result matches the expected result
-      testCase.verifyEqual(dqdtActual,dqdtExpected)
+         % Verify that the actual result matches the expected result
+         testCase.verifyEqual(dqdtActual,dqdtExpected)
 
-      % for testing:
-      % DerivMethod = 'CTS';
-      % [~,dqdtActual] = bfra.getdqdt(t,q,[],DerivMethod);
-      % isequal(dqdtActual,dqdtExpected)
-      % scatterfit(dqdtActual,dqdtExpected)
+         % for testing:
+         % DerivMethod = 'CTS';
+         % [~,dqdtActual] = bfra.getdqdt(t,q,[],DerivMethod);
+         % isequal(dqdtActual,dqdtExpected)
+         % scatterfit(dqdtActual,dqdtExpected)
 
       end
 
@@ -70,48 +70,48 @@ classdef ParameterizedTestBfra < matlab.unittest.TestCase
       %-------------------------------------------
       function test_fitab(testCase,RecessionExponent,FitMethod)
 
-      % define values to generate the test data
-      a = 1e-2;
-      q0 = 100;
-      t = 1:100;
-      b = RecessionExponent;
+         % define values to generate the test data
+         a = 1e-2;
+         q0 = 100;
+         t = 1:100;
+         b = RecessionExponent;
 
[…]
