### 2023-08-25 98499ac
feat(todatenum): add function to convert datetime to datenum consistently
