+++ +bfra/eventfinder.m
@@ -150,6 +150,9 @@ rmconvex    = parser.Results.rmconvex;
 rmnochange  = parser.Results.rmnochange;
 rmrain      = parser.Results.rmrain;
 
+% Convert datetime to double if datetime was passed in
+t = todatenum(t);
+
 validateattributes(t, {'double'},{'size', size(q)}, funcname,'t', 1)
 validateattributes(nmin, {'double'},{'>', 2}, funcname,'nmin')
 
