### 2023-08-25 d3a0fd9
fix(globalfit): fix input parser variable names
change drainableporos to drainableporosity, nreps to bootreps

change input "K" to "Results" to match fitevents
