+++ demos/bfra_demo_kuparuk.m
@@ -52,7 +52,7 @@ load('data/annualdata.mat','Data');
 opts.getevents = bfra.setopts('getevents');
 opts.fitevents = bfra.setopts('fitevents');
 opts.globalfit = bfra.setopts('globalfit','drainagearea',A,'aquiferdepth',D, ...
-   'streamlength',L,'drainagedens',Dd,'bootfit',bootfit,'nreps',nreps, ...
+   'streamlength',L,'drainagedens',Dd,'bootfit',bootfit,'bootreps',nreps, ...
    'plotfits',true);
 %% 
 % Get all recession events and then fit them
