### 2023-08-25 47a6329
feat(test-private-completions): enable private function tests
add function to return private function handle for use in tests

add completions function, use it to return list of private functions

add test of private todatenum function

update functionSignatures to use completions

move prepinput content out of wrapevents to private/preparecalendar
