### 2023-08-26 928d6e1
refactor(private)!: move +util, +deps, and non-user facing functions to private
replace prefixed function calls to newly private functions

this refactor enabled by bfra.privatefunction and associated test framework

remove or replace functions with user-facing interfaces e.g. taufunc replaced by call to getfunction
