+++ +bfra/privatefunction.m
@@ -1,9 +1,14 @@
 function funcHandle = privatefunction(funcName)
 
-   switch funcName
-      case 'todatenum'
-         funcHandle = @todatenum;
-      otherwise
-         error('Unknown function name.');
-   end
+   funcHandle = str2func(funcName);
+   
+   % use this to confirm correct scoping
+   % functions(funcHandle)
+   
+   % switch funcName
+   %    case 'todatenum'
+   %       funcHandle = @todatenum;
+   %    otherwise
+   %       error('Unknown function name.');
+   % end
 end
