### 2023-08-26 3c06f8a
refactor(inputParser): remove magicParser from project
update matfunclib deps and adhoc linting / parser updates
