+++ +bfra/+util/timetablereduce.m
@@ -4,17 +4,11 @@ function NewData = timetablereduce(Data,varargin)
 % NewData = timetablereduce(Data,'keeptime',true) keeps the time column in the
 % case of single vector input and returns the vector with new header 'mu'
 % 
-% See also stderror
+% See also: stderror
 
-%-------------------------------------------------------------------------------
-p = bfra.deps.magicParser;
-p.FunctionName = mfilename;
-p.addRequired( 'Data', @(x)istimetable(x));
-p.addParameter('alpha', 0.32, @(x)isnumeric(x));
-p.addParameter('keeptime', true, @(x)islogical(x));
-p.parseMagically('caller');
-alpha = p.Results.alpha; 
-%-------------------------------------------------------------------------------
+% parse inputs
+[Data, dim, alpha, keeptime] = parseinputs(Data, mfilename, varargin{:});
+
 Data.Properties.DimensionNames{1} = 'Time';
 Time = Data.Time;
 
@@ -34,15 +28,24 @@ if width(Data) == 1 && keeptime == true
    return
 end
 
-% compute mean, stderr, ci, etc
-[SE,CI,PM,mu,sigma] = bfra.util.stderror(table2array(Data),'alpha',alpha); 
+if dim == 2
+   % compute mean, stderr, ci, etc
+   [SE,CI,PM,mu,sigma] = bfra.util.stderr(table2array(Data),'alpha',alpha); 
+elseif dim == 1
+   [SE,CI,PM,mu,sigma] = bfra.util.stderr(transpose(table2array(Data)),'alpha',alpha); 
+end
+   
 CIL = CI(:,1); CIH = CI(:,2);
 
-if isscalar(mu)
+if isscalar(mu) || dim == 1
    Time = mean(Time);
 end
 
-NewData = timetable(Time,mu,sigma,SE,CIL,CIH,PM);
+if dim == 2
+   NewData = timetable(Time,mu,sigma,SE,CIL,CIH,PM);
+elseif dim == 1
+   NewData = table(mu,sigma,SE,CIL,CIH,PM,'RowNames',Data.Properties.VariableNames);
+end
 
 % copy over properties (might not work, if fails, fix brace indexing)
 if ~isempty(Data.Properties.VariableUnits)
@@ -52,3 +55,33 @@ elseif ~isempty(Data.Properties.VariableContinuity)
 end
 
 
+function [Data, dim, alpha, keeptime] = parseinputs(Data, funcname, varargin)
+
+p = inputParser;
+p.FunctionName = funcname;
+
+p.addRequired( 'Data', @(x) istimetable(x));
+p.addOptional( 'dim', 2, @(x) isnumericscalar(x));
+p.addParameter('alpha', 0.32, @(x) isnumeric(x));
+p.addParameter('keeptime', true, @(x) islogical(x));
+
+p.parse(Data,varargin{:});
+
+Data = p.Results.Data;
+dim = p.Results.dim;
+alpha = p.Results.alpha;
+keeptime = p.Results.keeptime;
+
+
+function T = renametimetabletimevar(T)
+%RENAMETIMETABLETIMEVAR rename the time variable in table T to 'Time'
+% 
+%  T = renametimetabletimevar(T)
+% 
+% See also:
+
+dims = T.Properties.DimensionNames;
+   
+if string(dims{1}) ~= "Time"
+   T.Properties.DimensionNames{1} = 'Time';
+end

[…]
