### 2023-08-27 0b9b00c
fix(demos): move mfiles to subfolder so live scripts are not shadowed
update new parameter names in calls to globalopts

move testdata function out of +test subpackage to avoid confusion

remove unused validationModule in prep for moving validation functions to private
