+++ demos/mfiles/bfra_demo_2.m
@@ -12,10 +12,7 @@
 % streamflow gage 1596000. The data are included in the |data/| folder as a sample
 % dataset with this toolbox, or can be downloaded from <https://waterdata.usgs.gov/monitoring-location/15896000.
 % https://waterdata.usgs.gov/monitoring-location/15896000.>
-%%
 %% Preparing data for recession analysis
-%
-%
 % Set the sitename.
 
 sitename = bfra.basinname('KUPARUK R NR DEADHORSE AK');
@@ -23,18 +20,14 @@ sitename = bfra.basinname('KUPARUK R NR DEADHORSE AK');
 % Load streamflow data for the test basin into the workspace. In the sample
 % dataset, the variable |T| is _time_, |Q| is _discharge_, and |R| is _rainfall_.
 
-[T,Q,R] = bfra.test.testdata('example');
+[T, Q, R] = bfra.testdata('example');
 %%
 % Set the basin area and geomorphological parameters.
 
 Dd = 0.8;            % drainage density
 A = 8.6545e9;        % basin area [m2]
 L = A*Dd/1000;       % active stream length
-D = 0.5;             % saturated aquifer thickness
-%%
 %% Set options for detecting and fitting recession events
-%
-%
 % The toolbox supports three main tasks: recession event detection, recession
 % event curve-fitting, and parameter estimation via distribution fitting. The
 % |setopts| function provides an interface to set options for the algorithms that
@@ -45,8 +38,6 @@ D = 0.5;             % saturated aquifer thickness
 opts.getevents = bfra.setopts('getevents');
 opts.fitevents = bfra.setopts('fitevents');
 %%
-%
-%
 % In addition to |getevents and| |fitevents|, in this example we set options
 % for |globalfit|. The term "globalfit" differentiates two fitting procedures:
 % one that fits individual recession events to estimate parameters specific to
@@ -56,26 +47,20 @@ opts.fitevents = bfra.setopts('fitevents');
 %
 % Set options for |globalfit.|
 
-opts.globalfit = bfra.setopts('globalfit','drainagearea',A,'streamlength',L, ...
-   'drainagedens',Dd,'aquiferdepth',D,'isflat',true,'plotfits',true);
-%%
+opts.globalfit = bfra.setopts('globalfit', ...
+   'drainagearea', A, 'streamlength', L, 'drainagedensity', Dd, ...
+   'isflat', true, 'plotfits', true);
 %% Recession analysis workflow
-%
-%
 % The following steps demonstrate a typical workflow in baseflow recession analysis.
 % The steps consist of getting a list of recession events, fitting the recession
 % events, and then estimating the recession parameters _a_ and _b_.
 % Step 1. Get events
 
-EventData = bfra.getevents(T,Q,R,opts.getevents);
-%
+EventData = bfra.getevents(T, Q, R, opts.getevents);
 % Step 2. Fit events
 
-[EventFits,FitResults] = bfra.fitevents(EventData,opts.fitevents);
-%
+[EventFits, FitsTable] = bfra.fitevents(EventData, opts.fitevents);
 % Step 3. Fit recession parameters _a_ and _b_
-%
-%
 % The |globalfit| function fits a Pareto distribution to the sample population
 % of event-scale parameters stored in |FitsTable|. The parameters of the Pareto
 % distribution $\tau_0$ and $\alpha$ can be related to the traditional recession
@@ -98,14 +83,9 @@ EventData = bfra.getevents(T,Q,R,opts.getevents);
 % recession analysis, identifying the "late-time" regime is important because
 % the parameters for this regime provide insight into aquifer properties that
 % control late-time drainage.
-%
-%
 
[…]

[…]
