### 2023-08-27 dbbb86d
fix(setlogticks): add method to handle zeros in ylims to avoid Inf
