+++ +bfra/hyetograph.m
@@ -1,131 +1,146 @@
-function H = hyetograph(time,flow,prec,varargin)
-% function H = hyetograph(time,flow,prec,t1,t2)
-%HYETOGRAPH Plot a discharge rainfall hyetograph
-%
-%  H = hyetograph(time,flow,ppt) plots hyetograph using data in flow and ppt
-%
-%  H = hyetograph(time,flow,ppt,t1,t2) plots hyetograph using data in flow and
-%  ppt for time period bounded by datetimes t1 and t2
-%
-% Example, how to later access the plot objects
-%  H = hyetograph(...);
-%
-% Find specific handles using 'Tag' or 'Name'
-%
-% fig = findobj(H, 'Name', 'Hyetograph');
-% ax_streamflow = findobj(H, 'Tag', 'StreamflowAxis');
-% ax_precipitation = findobj(H, 'Tag', 'PrecipitationAxis');
-%
-% See also
+function H = hyetograph(time, flow, prec, varargin)
+   % function H = hyetograph(time,flow,prec,t1,t2)
+   %HYETOGRAPH Plot a discharge rainfall hyetograph
+   %
+   %  H = hyetograph(time,flow,ppt) plots hyetograph using data in flow and ppt
+   %
+   %  H = hyetograph(time,flow,ppt,t1,t2) plots hyetograph using data in flow
+   %  and ppt for time period bounded by datetimes t1 and t2
+   %
+   % Example, how to later access the plot objects
+   %  H = hyetograph(...);
+   %
+   % Find specific handles using 'Tag' or 'Name'
+   %
+   % fig = findobj(H, 'Name', 'Hyetograph');
+   % ax_streamflow = findobj(H, 'Tag', 'StreamflowAxis');
+   % ax_precipitation = findobj(H, 'Tag', 'PrecipitationAxis');
+   %
+   % See also:
 
-% if called with no input, open this file
-if nargin == 0; open([mfilename('fullpath'),'.m']); return; end
+   % if called with no input, open this file
+   if nargin == 0; open([mfilename('fullpath'),'.m']); return; end
 
-% parse inputs
+   % parse inputs
+   [time, flow, prec, t1, t2, units, fig] = parseinputs(mfilename, time, flow, ...
+      prec, varargin{:});
 
-[time, flow, prec, t1, t2, units] = parseinputs(mfilename, time, flow, ...
-   prec, varargin{:});
+
+   % initialize container for output handles
+   if bfra.util.isoctave
+      H = zeros(4,1);
+   else
+      H = gobjects(4,1); % figure, ax1, plot1, plot2
+   end
+
+   % Set the figure handle tag
+   set(fig, 'Name', 'Hyetograph', 'Tag', 'HyetographFigure');
+
+   % convert to columns
+   time = time(:);
+   prec = prec(:);
+   flow = flow(:);
+
+   % convert to datetime
+   if ~bfra.util.isoctave
+      if ~isdatetime(time); time = datetime(time, 'ConvertFrom', 'datenum'); end
+      if ~isdatetime(t1); t1 = datetime(t1, 'ConvertFrom', 'datenum'); end
+      if ~isdatetime(t2); t2 = datetime(t2, 'ConvertFrom', 'datenum'); end
+
+      % trim to t1,t2 timespan
+      prec = prec(isbetween(time, t1, t2));
+      flow = flow(isbetween(time, t1, t2));
+      time = time(isbetween(time, t1, t2));
+   else
[…]
