### 2023-09-04 b8ad8db
refactor(getevents-eventplotter): Use cellfun methods to simplify loops over events.
Eliminate setEventEmpty except for getevents.

Refactor eventplotter: Add d2qdt parameter.

Use backward difference instead of central.

Add struct2vec call to plot all points first then i<type> points on top.
