+++ +bfra/getevents.m
@@ -1,187 +1,206 @@
 function [Events,Info] = getevents(T,Q,R,varargin)
-%GETEVENTS get individual recession events from daily timeseries T, Q, and R.
-%
-% Syntax
-%
-%     Events = getevents(T,Q,R,varargin)
-%
-% Description
-%
-%     Events = getevents(T,Q,R) Detects, processes, and organizes individual
-%     recession events from daily hydrograph timeseries T, Q, and rainfall R
-%     using default algorithm options. Event discharge, timestamps, and
-%     diagnostic info about the events are returned in output structure Events.
-%     Note: this function is a wrapper around eventfinder to perform pre- and
-%     post-processing and organize all recession events into the Events
-%     structure.
-%
-%     Events = getevents(___,opts) uses user-defined options. See bfra.setopts
-%     for default options and optional values.
-%
-%     [Events,Info] = getevents(___) also returns struct Info which contains the
-%     indices of the identified local maxima, minima, convex points, candidate
-%     recession values, kept recession values, and the start and stop index of
-%     each kept event. Use this information with bfra.eventplotter.
-%
-%     Tip: events are identified by their indices on the t,q,r arrays, so if
-%     any filtering is applied prior to passing in the arrays, the data needs
-%     to be used in subsequent functions or the indices won't be correct
-%
-% Required inputs
-%
-%     T     time, nx1 vector of datetimes
-%     Q     flow, nx1 vector of discharge (length/time) (assumed m3/day/day)
-%     R     rain, nx1 vector of rainfall (length/time) (assumed mm/day)
-%
-% Optional name-value inputs
-%
-%     opts        (optional) structure containing the following fields:
-%     qmin        minimum flow value, below which values are set nan
-%     nmin        minimum event length
-%     fmax        maximum # of missing values gap-filled
-%     rmax        maximum run of sequential constant values
-%     rmin        minimum rainfall required to censor flow (mm/day?)
-%     cmax        maximum run of sequential convex dQ/dt values
-%     rmconvex    remove convex derivatives
-%     rmnochange  remove consecutive constant derivates
-%     rmrain      remove rainfall
-%     pickevents  option to manually pick events
-%     plotevents  option to plot picked events
-%
-% See also fitevents, eventfinder, eventsplitter, eventpicker, eventplotter
-%
-% Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
-% 
-% Note to users: for data-heavy workflows, replace the Events output struct
-% with the eventTags list. The Events struct includes the input data and the
-% event-data for convenience, but the event-data can be easily extracted from
-% the input data using the eventTags.
+   %GETEVENTS Get recession events from daily hydrograph timeseries T, Q, and R.
+   %
+   %% Syntax
+   %
+   %     Events = getevents(T,Q,R)
+   %     [Events, Info] = getevents(T,Q,R,opts)
+   %     [Events, Info] = getevents(_, 'qmin', qmin)
+   %     [Events, Info] = getevents(_, 'nmin', nmin)
+   %     [Events, Info] = getevents(_, 'fmax', fmax)
+   %     [Events, Info] = getevents(_, 'rmax', rmax)
+   %     [Events, Info] = getevents(_, 'rmin', rmin)
+   %     [Events, Info] = getevents(_, 'cmax', cmax)
[…]
