+++ +bfra/eventplotter.m
@@ -1,186 +1,216 @@
 function [h,f] = eventplotter(T,Q,R,Info,varargin)
-%EVENTPLOTTER plot recession events detected by eventfinder
-%
-% Syntax
-%
-%     h = eventplotter(t,q,r,Info,varargin)
-%
-% Description
-%
-%     h = eventplotter(t,q,r,Info) Plots recession events identified by
-%     eventfinder on hydrograph t,q and rainfall r. Info is a structure returned
-%     by eventfinder that contains the indices of the start and end of each
-%     event as well as the local peaks, minimums, and runlength. An option to
-%     plot dq/dt as positive or negative values is available.
-%
-% Required inputs
-%
-%     T        time
-%     Q        flow (m3/time)
-%     R        rain (mm/time)
-%     Info     Info structure returned by findevents.m
-%
-% Optional name-value inputs
-%
-%  dqdt: user-provided dqdt, default = centered finite diff
-%  plotevents: logical, name-value e.g. 'plotevents',true
-%  plotneg: logical, name-value
-%
-% See also getevents, eventfinder, eventpicker, eventsplitter
-%
-% Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
+   %EVENTPLOTTER plot recession events detected by eventfinder
+   %
+   % Syntax
+   %
+   %     h = eventplotter(t,q,r,Info)
+   %     h = eventplotter(t,q,r,Info,eventTags)
+   %     h = eventplotter(_,'dqdt',dQdt)
+   %     h = eventplotter(_,'d2qdt',d2Qdt)
+   %     h = eventplotter(_,'reversesign',true)
+   %     h = eventplotter(_,'plotevents',false)
+   %
+   % Description
+   %
+   %     h = eventplotter(t,q,r,Info) Plots recession events identified by
+   %     eventfinder on hydrograph t,q and rainfall r. Info is a structure
+   %     returned by eventfinder that contains the indices of the start and end
+   %     of each event as well as the local peaks, minimums, and runlength. An
+   %     option to plot dq/dt as positive or negative values is available.
+   %
+   % Required inputs
+   %
+   %     T        time, numeric vector
+   %     Q        flow, numeric vector (m3/time)
+   %     R        rain, numeric vector (mm/time)
+   %     Info     Info, structure returned by getevents.m
+   %
+   % Optional inputs
+   % 
+   %     eventTags   numeric vector, indices for events to plot 
+   % 
+   % Optional name-value inputs
+   %
+   %     dqdt: user-provided first derivative, default = backward finite diff
+   %     d2qdt: user-provided second derivative, default = backward finite diff
+   %     plotevents: logical, name-value e.g. 'plotevents',false to return
+   %     without plotting (useful for calling from another function without
+   %     needing to update the syntax)
+   %     reversesign: logical, name-value indicating to plot dQ/dt instead of
+   %     -dQ/dt on the y-axis.
+   %     
+   %
+   % See also getevents, eventfinder, eventpicker, eventsplitter
+   %
+   % Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
 
-% if called with no input, open this file
-if nargin == 0; open(mfilename('fullpath')); return; end
+   % if called with no input, open this file
[…]
