### 2023-09-05 57d2068
feat(test_demos): add test_demo script and update demos to not clear testCase
