+++ +bfra/private/demopath.m
@@ -1,0 +1,3 @@
+function fullpath = demopath(varargin)
+   fullpath = fullfile(basepath(), 'demos', varargin{:});
+end
