### 2023-09-05 082549c
refactor(eventfinder): replace loop with arrayfun to simplify function.
Move non nan segments to function.

Intermediate commit before moving eventsplitter to local function of eventfinder.
