+++ +bfra/eventfinder.m
@@ -51,30 +51,19 @@ function [T,Q,R,Info] = eventfinder(t,q,r,varargin)
    [t, q, r, nmin, fmax, rmax, rmin, rmconvex, rmnochange, rmrain] = ...
       parseinputs(t, q, r, mfilename, varargin{:});
 
-   % MAIN CODE
-   iS = unique([1; find(diff(isnan(q)) == -1)+1]);    % start non-nan segments
-   iE = unique([find(diff(isnan(q)) == 1);numel(q)]); % end non-nan segments
-   L  = iE - iS + 1;                                  % segment lengths
-
-   % Remove events shorter than nmin
-   iS = iS(L >= nmin);
-   iE = iE(L >= nmin);
-
-   % initialize the combined events
-   [T, Q, R, Info] = deal(cell(numel(iS), 1));
+   % Find recession events within NaN-separated flow segments.
+   [iS, iE] = nonnansegments(q, nmin);
          
-   for n = 1:numel(iS)
-
-      [T{n}, Q{n}, R{n}, Info{n}] = bfra.eventsplitter( ...
-         t(iS(n):iE(n)), q(iS(n):iE(n)), r(iS(n):iE(n)), ...
-         'nmin', nmin, ...
-         'fmax', fmax, ...
-         'rmax', rmax, ...
-         'rmin', rmin, ...
-         'rmrain', rmrain, ...
-         'rmconvex', rmconvex, ...
-         'rmnochange', rmnochange);
-   end
+   [T, Q, R, Info] = arrayfun(@(s, e) bfra.eventsplitter( ...
+      t(s:e), q(s:e), r(s:e), ...
+      'nmin', nmin, ...
+      'fmax', fmax, ...
+      'rmax', rmax, ...
+      'rmin', rmin, ...
+      'rmrain', rmrain, ...
+      'rmconvex', rmconvex, ...
+      'rmnochange', rmnochange), ...
+      iS, iE, 'UniformOutput',false);
    
    if isempty(T)
       [T, Q, R, Info] = setEventEmpty();
