+++ +bfra/eventsplitter.m
@@ -1,234 +1,239 @@
 function [T,Q,R,Info] = eventsplitter(t,q,r,varargin)
-%eventsplitter split detected recession events into useable segments for fitting
-%
-% Syntax
-%
-%     [T,Q,R,Info] = eventsplitter(t,q,r,varargin)
-%
-% Description
-%
-%     Split recession events detected by eventfinder into individual segments
-%     ready to fit with bfra.fitab or bfra.fitevents. For example, if an event
-%     is interrupted by rainfall or if rainfall is detected from convex dq/dt,
-%     the event can be split into separate segments thought to represent
-%     uninterrupted baseflow.
-%
-% Required inputs
-%
-%     t        time
-%     q        flow (m3/time)
-%     r        rain (mm/time)
-%
-% Optional name-value inputs
-%
-%     nmin        minimum event length
-%     fmax        maximum # of missing values gap-filled
-%     rmax        maximum run of sequential constant values
-%     rmin        minimum rainfall required to censor flow (mm/day?)
-%     rmconvex    remove convex derivatives
-%     rmnochange  remove consecutive constant derivates
-%     rmrain      remove rainfall
-%
-% See also getevents, eventfinder, eventpicker, eventplotter
-%
-% Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
+   %eventsplitter Split detected recession events into segments for fitting.
+   %
+   % Syntax
+   %
+   %     [T,Q,R,Info] = eventsplitter(t,q,r,varargin)
+   %
+   % Description
+   %
+   %     Split recession events detected by eventfinder into individual segments
+   %     ready to fit with bfra.fitab or bfra.fitevents. For example, if an
+   %     event is interrupted by rainfall or if rainfall is detected from convex
+   %     dq/dt, the event can be split into separate segments thought to
+   %     represent uninterrupted baseflow.
+   %
+   % Required inputs
+   %
+   %     t        time
+   %     q        flow (m3/time)
+   %     r        rain (mm/time)
+   %
+   % Optional name-value inputs
+   %
+   %     nmin        minimum event length
+   %     fmax        maximum # of missing values gap-filled
+   %     rmax        maximum run of sequential constant values
+   %     rmin        minimum rainfall required to censor flow (mm/day?)
+   %     rmconvex    remove convex derivatives
+   %     rmnochange  remove consecutive constant derivates
+   %     rmrain      remove rainfall
+   %
+   % See also getevents, eventfinder, eventpicker, eventplotter
+   %
+   % Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
 
-% if called with no input, open this file
-if nargin == 0; open(mfilename('fullpath')); return; end
+   % if called with no input, open this file
+   if nargin == 0; open(mfilename('fullpath')); return; end
 
-persistent inoctave
-if isempty(inoctave); inoctave = exist("OCTAVE_VERSION", "builtin")>0;
-end
-
-% if nmin is set to 0 (and maybe if it is set to 1) this method will fail
-% because runlength returns 1 for consecutive nan values, see isminlength.
-
-% PARSE INPUTS
[…]
