### 2023-09-05 90d1d72
fix(eventsplitter): use [-3 ... 3] rather than -3:3
