### 2023-09-05 19275b0
refactor(eventfinder): move eventsplitter to local function of eventfinder
