+++ +bfra/eventfinder.m
@@ -1,4 +1,4 @@
-function [T,Q,R,Info] = eventfinder(t,q,r,varargin)
+function [T, Q, R, Info] = eventfinder(t, q, r, varargin)
    %EVENTFINDER find recession events using hydrograph and rainfall timeseries.
    %
    % Syntax
@@ -48,28 +48,19 @@ function [T,Q,R,Info] = eventfinder(t,q,r,varargin)
    if nargin == 0; open(mfilename('fullpath')); return; end
 
    % PARSE INPUTS
-   [t, q, r, nmin, fmax, rmax, rmin, rmconvex, rmnochange, rmrain] = ...
-      parseinputs(t, q, r, mfilename, varargin{:});
+   [t, q, r, opts] = parseinputs(t, q, r, mfilename, varargin{:});
 
-   % Find recession events within NaN-separated flow segments.
-   [iS, iE] = nonnansegments(q, nmin);
-         
-   [T, Q, R, Info] = arrayfun(@(s, e) bfra.eventsplitter( ...
-      t(s:e), q(s:e), r(s:e), ...
-      'nmin', nmin, ...
-      'fmax', fmax, ...
-      'rmax', rmax, ...
-      'rmin', rmin, ...
-      'rmrain', rmrain, ...
-      'rmconvex', rmconvex, ...
-      'rmnochange', rmnochange), ...
-      iS, iE, 'UniformOutput',false);
-   
+   % MAIN CODE
+   [iS, iE] = nonnansegments(q, opts.nmin);
+
+   [T, Q, R, Info] = arrayfun(@(s, e) ...
+      eventsplitter(t(s:e), q(s:e), r(s:e), opts), iS, iE, 'Uniform', false);
+
    if isempty(T)
       [T, Q, R, Info] = setEventEmpty();
       return
    end
-   
+
    % iS is the start of the entire flow segment, but the indices in Info are
    % relative to the event returned by eventsplitter. Add iS to these indices,
    fields = fieldnames(Info{1});
@@ -78,15 +69,193 @@ function [T,Q,R,Info] = eventfinder(t,q,r,varargin)
          Info{n}.(fields{m}) = Info{n}.(fields{m}) + iS(n) - 1;
       end
    end
-   T = vertcat(T{:});
-   Q = vertcat(Q{:});
-   R = vertcat(R{:});
+   [T, Q, R] = deal(vertcat(T{:}), vertcat(Q{:}), vertcat(R{:}));
    Info = catstructfields(1, Info{:});
 end
 
+function [T, Q, R, Info] = eventsplitter(t, q, r, opts)
+   %eventsplitter Split detected recession events into segments for fitting.
+   %
+   % Syntax
+   %
+   %     [T,Q,R,Info] = eventsplitter(t,q,r,opts)
+   %
+   % Description
+   %
+   %     Split recession events detected by eventfinder into individual segments
+   %     ready to fit with bfra.fitab or bfra.fitevents. For example, if an
+   %     event is interrupted by rainfall or if rainfall is detected from convex
+   %     dq/dt, the event can be split into separate segments thought to
+   %     represent uninterrupted baseflow. Follows recommendations in Dralle et
+   %     al. 2017.
+   %
+   % Required inputs
+   %
+   %     t           time
+   %     q           flow (m3/time)
+   %     r           rain (mm/time)
+   %     opts        structure containing the following fields:
+   %     nmin        minimum event length
+   %     fmax        maximum # of missing values gap-filled
+   %     rmax        maximum run of sequential constant values
+   %     rmin        minimum rainfall required to censor flow (mm/day?)
[…]
