+++ +bfra/eventsplitter.m
@@ -1,239 +1,0 @@
-function [T,Q,R,Info] = eventsplitter(t,q,r,varargin)
-   %eventsplitter Split detected recession events into segments for fitting.
-   %
-   % Syntax
-   %
-   %     [T,Q,R,Info] = eventsplitter(t,q,r,varargin)
-   %
-   % Description
-   %
-   %     Split recession events detected by eventfinder into individual segments
-   %     ready to fit with bfra.fitab or bfra.fitevents. For example, if an
-   %     event is interrupted by rainfall or if rainfall is detected from convex
-   %     dq/dt, the event can be split into separate segments thought to
-   %     represent uninterrupted baseflow.
-   %
-   % Required inputs
-   %
-   %     t        time
-   %     q        flow (m3/time)
-   %     r        rain (mm/time)
-   %
-   % Optional name-value inputs
-   %
-   %     nmin        minimum event length
-   %     fmax        maximum # of missing values gap-filled
-   %     rmax        maximum run of sequential constant values
-   %     rmin        minimum rainfall required to censor flow (mm/day?)
-   %     rmconvex    remove convex derivatives
-   %     rmnochange  remove consecutive constant derivates
-   %     rmrain      remove rainfall
-   %
-   % See also getevents, eventfinder, eventpicker, eventplotter
-   %
-   % Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
-
-   % if called with no input, open this file
-   if nargin == 0; open(mfilename('fullpath')); return; end
-
-   persistent inoctave
-   if isempty(inoctave); inoctave = exist("OCTAVE_VERSION", "builtin")>0;
-   end
-
-   % if nmin is set to 0 (and maybe if it is set to 1) this method will fail
-   % because runlength returns 1 for consecutive nan values, see isminlength.
-
-   % PARSE INPUTS
-   [t, q, r, nmin, ~, rmax, rmin, rmconvex, rmnochange, rmrain] = parseinputs( ...
-      t, q, r, mfilename, varargin{:});
-   debug = false;
-
-   % MAIN CODE (follows recommendations in Dralle et al. 2017)
-
-   % get a 3-day smoothed timeseries to control false positive convexity
-   if inoctave
-      qS = nanmovmean(q,3);
-   else
-      qS = movmean(q, 3, 'omitnan');
-   end
-
-   % First and second differences of the discharge and the smoothed discharge
-   dqdt = [0; diff(q)];
-   dqdtS = [0; diff(qS)];
-   d2qdt = [0; diff(dqdt)];
-   d2qdtS = [0; diff(dqdtS)];
-
-   % Part 1: find data to be excluded (run the filters)
-
-   % filters: positive dq/dt, peaks +1 day, convex +1 day, and minima -1 day
-   ipos = find(dqdt > 0); % increasing flow (positive dq/dt)
-   imax = find(islocalmax(q)); % local maxima
-   imin = find(islocalmin(q)); % local minima
-   icon = find(islocalmax(dqdt) & islocalmax(dqdtS)) + 1;  % convex + 1 day
-
-   % This also finds the transitions from convex up to concave up
-   % icon = find( (islocalmax(dqdt) & islocalmax(dqdtS)) | ...
-   %   (islocalmin(dqdt) & islocalmin(dqdtS)) ) + 1;  % convex + 1 day
-
-   % Dralle: keep if dq/dt<0 AND d2q/dt2>0 (concave up) in BOTH the raw and
-   % smoothed data, i.e. exclude if both raw AND smoothed data are convex
[…]
