### 2023-09-05 75618df
chore: lint private functions
