+++ +bfra/private/emptyaxes.m
@@ -1,8 +1,8 @@
 function ax = emptyaxes
-
-if isoctave
-   ax = [];
-   % ax = get(gcf, "currentaxes");
-else
-   ax = gobjects(0);
+   if isoctave
+      ax = [];
+      % ax = get(gcf, "currentaxes");
+   else
+      ax = gobjects(0);
+   end
 end
