+++ +bfra/private/isdoublescalar.m
@@ -1,2 +1,3 @@
 function tf = isdoublescalar(x)
-tf = isa(x,'double') && isscalar(x);
+   tf = isa(x,'double') && isscalar(x);
+end
