+++ +bfra/private/setnan.m
@@ -1,137 +1,139 @@
 function [ DataOut ] = setnan(Data,varargin)
-%SETNAN sets logical indices in Data to nan
-%
-%  [ DataOut ] = setnan(Data) sets all values in Data nan
-% 
-%  [ DataOut ] = setnan(Data,nanval) sets values in Data that equal nanval nan
-% 
-%  [ DataOut ] = setnan(Data,[],naninds) sets naninds indices of Data nan
-% 
-% See also: rmnan, replacevars
+   %SETNAN sets logical indices in Data to nan
+   %
+   %  [ DataOut ] = setnan(Data) sets all values in Data nan
+   %
+   %  [ DataOut ] = setnan(Data,nanval) sets values in Data that equal nanval nan
+   %
+   %  [ DataOut ] = setnan(Data,[],naninds) sets naninds indices of Data nan
+   %
+   % See also: rmnan, replacevars
 
-% parse inputs
-[Data, nanval, naninds] = parseinputs(Data, mfilename, varargin{:});
+   % parse inputs
+   [Data, nanval, naninds] = parseinputs(Data, mfilename, varargin{:});
 
-% Note: below I detect which columns are numeric but i don't convert
-% from cell to numeric so if a table is imported with mixed data and it
-% should all be numeric, this won't work as expected (e.g. when reading
-% in new data for the first time and setting a known value nan)
+   % Note: below I detect which columns are numeric but i don't convert
+   % from cell to numeric so if a table is imported with mixed data and it
+   % should all be numeric, this won't work as expected (e.g. when reading
+   % in new data for the first time and setting a known value nan)
 
-wastable = istable(Data);           % was a table
-wastimetable = istimetable(Data);   % was a timetable
+   wastable = istable(Data);           % was a table
+   wastimetable = istimetable(Data);   % was a timetable
 
-% if only the Data was passed in, set all values nan
-if nargin == 1
-   DataOut = nan(size(Data));
-   if wastable || wastimetable
-      DataOut = replacevars(Data, Data.Properties.VariableNames, DataOut);
-      return
+   % if only the Data was passed in, set all values nan
+   if nargin == 1
+      DataOut = nan(size(Data));
+      if wastable || wastimetable
+         DataOut = replacevars(Data, Data.Properties.VariableNames, DataOut);
+         return
+      end
    end
-end
 
+   % if a table was passed in, prep it for replacement
+   if wastable
 
-% if a table was passed in, prep it for replacement
-if wastable
-   
-   inumeric = cellfun(@isnumeric, table2cell(Data(1, :)));
-   DataOut = table2array(Data(:, inumeric));
-   props = Data.Properties;
-   
-elseif wastimetable
-   
-   Time = Data.Time;
-   props = Data.Properties;    % get props before converting
-   iTime = Data.Properties.DimensionNames;
-   
-   Data = timetable2table(Data);
-   Data = Data(:, 2:end);      % remove time column
-   
-   inumeric = cellfun(@isnumeric, table2cell(Data(1, :)));
-   DataOut = table2array(Data(:, inumeric));
-   % i was gonna use iTime to remove the time column but only works if
-   % it is always called 'Time'
-else
-   DataOut = Data;
-end
+      inumeric = cellfun(@isnumeric, table2cell(Data(1, :)));
+      DataOut = table2array(Data(:, inumeric));
+      props = Data.Properties;
 
-% determine if nanval or naninds will be used, if the latter, assign naninds to
-% nanval b/c the final part below uses nanval for everything
-useval = false;
-useinds = false;
[…]
