+++ +bfra/private/istablelike.m
@@ -1,4 +1,4 @@
 function tf = istablelike(T)
-%ISTABLELIKE returns true if input T is a table or timetable. will be deprecated
-%on upgrade to R2021 or whichever release has istabular
-tf = istable(T) | istimetable(T);
+   %ISTABLELIKE Return true if input T is a table or timetable.
+   tf = istable(T) | istimetable(T);
+end
