+++ +bfra/private/basepath.m
@@ -1,8 +1,9 @@
 function projpath = basepath(varargin)
 
-projpath = fileparts(fileparts(fileparts(mfilename('fullpath'))));
+   projpath = fileparts(fileparts(fileparts(mfilename('fullpath'))));
 
-if nargin == 1
-   projpath = fullfile(projpath, varargin{:});
+   if nargin == 1
+      projpath = fullfile(projpath, varargin{:});
+   end
+
 end
-
