+++ +bfra/private/stderror.m
@@ -24,7 +24,7 @@ function [SE,CI,PM,mu,sig] = stderror(data,varargin)
    % parse inputs
    [data, dim, alpha] = parseinputs(data, mfilename, varargin{:});
 
-   % assume data is oriented columnwise, and we want the mean +/- stderr of each row
+   % assume data is oriented columnwise, and we want the mean +/- stderr of each row 
 
    % check for vector vs matrix input
    [r,c,p] = size(data);
