+++ +bfra/private/isaxis.m
@@ -1,7 +1,8 @@
 function tf = isaxis(ax)
-%ISAXIS logical check if axis object
-if isoctave
-   tf = isaxes(ax);
-else
-   tf = isa(ax,'matlab.graphics.axis.Axes');
+   %ISAXIS logical check if axis object
+   if isoctave
+      tf = isaxes(ax);
+   else
+      tf = isa(ax,'matlab.graphics.axis.Axes');
+   end
 end

[…]
