### 2023-09-05 2bc305b
chore: remove paper and compile workflow
