+++ paper.md
@@ -1,102 +1,0 @@
----
-title: 'baseflow: a MATLAB and GNU Octave package for baseflow recession analysis'
-tags:
-  - Matlab
-  - Octave
-  - hydrology
-  - groundwater
-  - baseflow
-  - streamflow
-  - aquifer
-  - Pareto distribution
-  - power law
-  - plfit
-authors:
-  - name: Matthew G. Cooper
-    orcid: 0000-0002-0165-209X
-    corresponding: true
-    affiliation: 1
-  - name: Tian Zhou
-    orcid: 0000-0003-1582-4005
-    affiliation: 1
-affiliations:
- - name: Atmospheric Science and Global Change, Pacific Northwest National Laboratory, Richland, WA, USA
-   index: 1
-date: today
-bibliography: paper.bib
----
-
-<!-- TODO: more detail on Figure 1 caption, let tian know about the review criteria section that says to compare with other software -->
-
-# Summary
-<!-- Begin your paper with a summary of the high-level functionality of your software for a non-specialist reader. Avoid jargon in this section -->
-`baseflow` is a MATLAB&reg; [@matlab_2020_] toolbox that facilitates baseflow recession analysis, a set of methods used in hydrologic science to infer aquifer properties that cannot be measured directly [@brutsaert_1977_WRR]. The widespread availability of streamflow measurements means that baseflow recession analysis can be used to estimate aquifer properties and their variations globally over the modern instrumental stream gage record. This toolbox offers easy-to-use functions for baseflow recession analysis using measured values of streamflow recorded on a daily timestep. It is primarily designed for shallow (depth << breadth) unconfined riparian aquifers that discharge groundwater laterally into adjacent stream channels, rather than confined aquifers or complex geologic structures which require specialized numerical groundwater models. The toolbox can also be used to study the collective behavior of individual hillslope aquifers that constitute hydrologic catchments, known as "watersheds", using a non-linear dynamical systems perspective [@kirchner_2009_WRR]. By leveraging recent advancements in the baseflow recession analysis literature [@roques_2017_AiWR; @dralle_2017_HESS], `baseflow` enables consistent, objective estimation of unconfined hillslope aquifer properties and conceptual catchment-scale aquifer properties.
-
-# Statement of need
-
-Baseflow is a vital component of streamflow that originates from groundwater sources rather than rainfall, surface water, or managed reservoir release [@hall_1968_WRR]. Baseflow plays a crucial role in sustaining water availability during dry seasons, particularly in systems lacking surface water storage [@cooper_2018_WRR]. The shape of the baseflow recession curve can provide insights into the physical properties of aquifers, which are geological structures that store and release groundwater to rivers [@brutsaert_1998_WRR; @troch_2013_WRR]. However, inferring the presence and amount of baseflow is a challenging task as it requires specialized algorithms involving signal processing, curve-fitting, and parameter estimation, using measured values of streamflow that are often contaminated by measurement error [@dralle_2017_HESS].
-
-\bigskip
-\begin{figure}[!tbh]
-\centering
-\includegraphics[width=\textwidth]{docs/figure_1.png}
-\caption{Example of baseflow recession analysis, the core functionality of the `baseflow` toolbox.}
-\label{fig:figure_1}
-\end{figure}
-
-<!-- ![Example of baseflow recession analysis, the core functionality of the `baseflow` toolbox.\label{fig:figure_1}](./docs/figure_1.png) -->
-
[…]
