+++ paper.bib
@@ -1,253 +1,0 @@
-@article{arciniega-esparza_2017_C&G,
-  title = {{{HYDRORECESSION}}: {{A Matlab}} Toolbox for Streamflow Recession Analysis},
-  shorttitle = {{{HYDRORECESSION}}},
-  author = {{Arciniega-Esparza}, Sa{\'u}l and {Bre{\~n}a-Naranjo}, Jos{\'e} Agust{\'i}n and {Pedrozo-Acu{\~n}a}, Adri{\'a}n and Appendini, Christian Mario},
-  year = {2017},
-  month = jan,
-  journal = {Computers \& Geosciences},
-  volume = {98},
-  pages = {87--92},
-  issn = {0098-3004},
-  doi = {10.1016/j.cageo.2016.10.005},
-  urldate = {2023-02-14},
-  langid = {english}
-}
-
-@misc{arciniega-esparza_2018_,
-  title = {{{HYDRORECESSION}}},
-  author = {{Arciniega-Esparza}, Sa{\'u}l},
-  year = {2018},
-  address = {{Natick, Massachusetts}},
-  urldate = {2023-02-11},
-  howpublished = {MATLAB Central File Exchange, The MathWorks Inc.}
-}
-
-@article{brutsaert_1977_WRR,
-  title = {Regionalized Drought Flow Hydrographs from a Mature Glaciated Plateau},
-  author = {Brutsaert, Wilfried and Nieber, John L.},
-  year = {1977},
-  month = jun,
-  journal = {Water Resources Research},
-  volume = {13},
-  number = {3},
-  pages = {637--643},
-  issn = {1944-7973},
-  doi = {10.1029/WR013i003p00637},
-  urldate = {2017-04-17},
-  langid = {english}
-}
-
-@article{brutsaert_1998_WRR,
-  title = {Basin-Scale Geohydrologic Drought Flow Features of Riparian Aquifers in the {{Southern Great Plains}}},
-  author = {Brutsaert, Wilfried and Lopez, James P.},
-  year = {1998},
-  month = feb,
-  journal = {Water Resources Research},
-  volume = {34},
-  number = {2},
-  pages = {233--240},
-  issn = {1944-7973},
-  doi = {10.1029/97WR03068},
-  urldate = {2017-05-03},
-  langid = {english}
-}
-
-@article{clauset_2009_SR,
-  title = {Power-{{Law Distributions}} in {{Empirical Data}}},
-  author = {Clauset, Aaron and Shalizi, Cosma Rohilla and Newman, M. E. J.},
-  year = {2009},
-  month = nov,
-  journal = {SIAM Review},
-  volume = {51},
-  number = {4},
-  pages = {661--703},
-  publisher = {{Society for Industrial and Applied Mathematics}},
-  issn = {0036-1445},
-  doi = {10.1137/070710111},
-  urldate = {2021-05-18}
-}
-
-@article{cooper_2018_WRR,
-  title = {Climate {{Elasticity}} of {{Low Flows}} in the {{Maritime Western U}}.{{S}}. {{Mountains}}},
-  author = {Cooper, M. G. and Schaperow, J. R. and Cooley, S. W. and Alam, S. and Smith, L. C. and Lettenmaier, D. P.},
-  year = {2018},
-  journal = {Water Resources Research},
-  volume = {54},
-  number = {8},
-  pages = {5602--5619},
-  issn = {1944-7973},
-  doi = {10.1029/2018WR022816},
-  urldate = {2019-03-05},
-  copyright = {\textcopyright 2018. American Geophysical Union. All Rights Reserved.},
-  langid = {english}
-}
-
-@article{cooper_2023_WRR,
[…]
