+++ .github/workflows/draft-pdf.yml
@@ -1,43 +1,0 @@
-# - name: Open Journals PDF Generator
-#   uses: openjournals/openjournals-draft-action@v.1.0
-
-on:
-  # Triggers the workflow on push or pull requests for the specified branch
-  push:
-    branches: [ joss ]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-jobs:
-  # This workflow contains a single job called "paper"
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - name: Checkout
-        uses: actions/checkout@v3
-      # Builds the pdf
-      # - name: Open Journals PDF Generator
-      #   uses: openjournals/openjournals-draft-action@v.1.0
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@v.1.0
-        # uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v1
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input paper.md
-          path: paper.pdf
-      # can test this at some point
-      # - name: save pdf to repo
-      #   uses: stefanzweifel/git-auto-commit-action@v4
-      #   with:
-      #   commit_message: Saved new PDF of paper
