+++ paper.bib
@@ -222,6 +222,21 @@
   langid = {english}
 }
 
+@article{rupp_2006_AiWR,
+  title = {Information, Artifacts, and Noise in {{dQ}}/Dt - {{Q}} Recession Analysis},
+  author = {Rupp, David E. and Selker, John S.},
+  year = {2006},
+  month = feb,
+  journal = {Advances in Water Resources},
+  series = {Experimental {{Hydrology}}: {{A Bright Future}}},
+  volume = {29},
+  number = {2},
+  pages = {154--160},
+  issn = {0309-1708},
+  doi = {10.1016/j.advwatres.2005.03.019},
+  urldate = {2017-05-16}
+}
+
 @article{rupp_2006_WRR,
   title = {On the Use of the {{Boussinesq}} Equation for Interpreting Recession Hydrographs from Sloping Aquifers},
   author = {Rupp, David E. and Selker, John S.},
@@ -234,6 +249,20 @@
   issn = {1944-7973},
   doi = {10.1029/2006WR005080},
   urldate = {2017-05-16},
+  langid = {english}
+}
+
+@article{troch_1993_WRR,
+  title = {Effective Water Table Depth to Describe Initial Conditions Prior to Storm Rainfall in Humid Regions},
+  author = {Troch, Peter A. and Troch, Fran{\c c}ois P. De and Brutsaert, Wilfried},
+  year = {1993},
+  journal = {Water Resources Research},
+  volume = {29},
+  number = {2},
+  pages = {427--434},
+  issn = {1944-7973},
+  doi = {10.1029/92WR02087},
+  urldate = {2021-07-09},
   langid = {english}
 }
 
