### 2023-09-06 4a6e36d
docs: revise paper
