+++ paper.md
@@ -26,52 +26,44 @@ date: today
 bibliography: paper.bib
 ---
 
-<!-- TODO: more detail on Figure 1 caption, let tian know about the review criteria section that says to compare with other software -->
-
 # Summary
-<!-- Begin your paper with a summary of the high-level functionality of your software for a non-specialist reader. Avoid jargon in this section -->
-`baseflow` is a MATLAB&reg; [@matlab_2020_] toolbox that facilitates baseflow recession analysis, a set of methods used in hydrologic science to infer aquifer properties that cannot be measured directly [@brutsaert_1977_WRR]. The widespread availability of streamflow measurements means that baseflow recession analysis can be used to estimate aquifer properties, such as hydraulic conductivity or drainable porosity, and their variations globally over the modern instrumental stream gage record. This toolbox offers easy-to-use functions for baseflow recession analysis using measured values of streamflow recorded on a daily timestep. It is primarily designed for shallow (depth << breadth) unconfined riparian aquifers that discharge groundwater laterally into adjacent stream channels, rather than confined aquifers or complex geologic structures which require specialized numerical groundwater models. The toolbox can also be used to study the collective behavior of individual hillslope aquifers that constitute hydrologic catchments, known as "watersheds", using a non-linear dynamical systems perspective [@kirchner_2009_WRR]. By leveraging recent advancements in the baseflow recession analysis literature [@roques_2017_AiWR; @dralle_2017_HESS], `baseflow` enables consistent, objective estimation of unconfined hillslope aquifer properties and conceptual catchment-scale aquifer properties.
+
+`baseflow` is a MATLAB&reg; [@matlab_2020_] toolbox designed for baseflow recession analysis, a technique used in hydrologic science to infer aquifer properties from streamflow [@brutsaert_1977_WRR]. Leveraging widely-available streamflow data, the toolbox can be used to estimate aquifer properties such as hydraulic conductivity and drainable porosity over the modern instrumental stream gage record. The toolbox is intended for analysis of measured streamflow values recorded on a daily timestep, and is tailored for shallow, unconfined riparian aquifers that discharge groundwater laterally into adjacent streams. Additionally, `baseflow` can analyze the collective behavior of individual hillslope aquifers constituting hydrologic catchments, known as "watersheds", from a non-linear dynamical systems perspective [@kirchner_2009_WRR]. The toolbox incorporates recent advances in baseflow recession analysis [@roques_2017_AiWR; @dralle_2017_HESS; @cooper_2023_WRR] to enable consistent, objective estimation of aquifer properties at both hillslope and catchment scales.
 
 # Statement of need
 
-Baseflow is the component of streamflow that originates solely from aquifer discharge, particularly during periods of no recharge [@hall_1968_WRR]. This excludes direct contributions from other sources like rainfall, surface water, or managed reservoir release, and essentially reflects the delayed release of groundwater storage into streams. Baseflow plays a crucial role in sustaining water availability during dry seasons, especially in regions with limited surface water storage [@cooper_2018_WRR]. Baseflow is therefore important for managing water scarcity, and understanding the baseflow recession process is a central task in hydrologic science [@hall_1968_WRR; @kirchner_2009_WRR; @troch_2013_WRR].
[…]
