### 2023-09-06 9735821
docs: add compile script and recompile paper
