+++ paper.md
@@ -1,76 +1,0 @@
----
-title: 'baseflow: a MATLAB and GNU Octave package for baseflow recession analysis'
-tags:
-  - Matlab
-  - Octave
-  - hydrology
-  - groundwater
-  - baseflow
-  - streamflow
-  - aquifer
-  - Pareto distribution
-  - power law
-  - plfit
-authors:
-  - name: Matthew G. Cooper
-    orcid: 0000-0002-0165-209X
-    corresponding: true
-    affiliation: 1
-  - name: Tian Zhou
-    orcid: 0000-0003-1582-4005
-    affiliation: 1
-affiliations:
- - name: Atmospheric Science and Global Change, Pacific Northwest National Laboratory, Richland, WA, USA
-   index: 1
-date: today
-bibliography: paper.bib
----
-
-# Summary
-
-`baseflow` is a MATLAB&reg; [@matlab_2020_] toolbox designed for baseflow recession analysis, a technique used in hydrologic science to infer aquifer properties from streamflow [@brutsaert_1977_WRR]. Leveraging widely-available streamflow data, the toolbox can be used to estimate aquifer properties such as hydraulic conductivity and drainable porosity over the modern instrumental stream gage record. The toolbox is intended for analysis of measured streamflow values recorded on a daily timestep, and is tailored for shallow, unconfined riparian aquifers that discharge groundwater laterally into adjacent streams. Additionally, `baseflow` can analyze the collective behavior of individual hillslope aquifers constituting hydrologic catchments, known as "watersheds", from a non-linear dynamical systems perspective [@kirchner_2009_WRR]. The toolbox incorporates recent advances in baseflow recession analysis [@roques_2017_AiWR; @dralle_2017_HESS; @cooper_2023_WRR] to enable consistent, objective estimation of aquifer properties at both hillslope and catchment scales.
-
-# Statement of need
-
-Baseflow is a vital component of streamflow sourced exclusively from aquifer discharge, rather than rainfall, surface water, or managed reservoir release [@hall_1968_WRR]. Baseflow essentially reflects the delayed release of groundwater storage into streams, and sustains water availability during dry seasons, especially in regions with limited surface water storage [@cooper_2018_WRR]. Baseflow is therefore important for managing water scarcity, and understanding the baseflow recession process is a central task in hydrologic science [@hall_1968_WRR; @kirchner_2009_WRR; @troch_2013_WRR].
-
-The baseflow recession curve captures the decline in streamflow during low-flow periods [@brutsaert_1977_WRR]. Parameters derived from this curve can be used to estimate unconfined aquifer properties, including hydraulic conductivity, drainable porosity, and the saturated zone thickness [@brutsaert_1998_WRR; @troch_1993_WRR; @rupp_2006_WRR]. Detecting baseflow segments from measured streamflow data, however, is a complex task that requires signal processing, curve fitting, and parameter estimation algorithms, further complicated by measurement error and noise [@dralle_2017_HESS; @rupp_2006_AiWR].
-
-\bigskip
-\begin{figure}[!tbh]
-\centering
-\includegraphics[width=\textwidth]{docs/figure_1.png}
-\caption{Example of baseflow recession analysis using the `baseflow` toolbox. (a) Daily discharge hydrograph for the Kuparuk River, Alaska recorded at United States Geological Survey gage 1596000 and one detected recession event. (b-c) Observed and fitted discharge values ($Q$) and their rate of change ($-\mathrm{d}Q/\mathrm{d}t$) for the event highlighted in (a). (d) Log-log relationship between $Q$ and $-\mathrm{d}Q/\mathrm{d}t$, comparing three fitting methods to estimate parameters $a$ and $b$: ordinary least squares (ols), nonlinear least squares (nls), and quantile regression (qtl).}
[…]
