+++ paper/paper.bib
@@ -185,10 +185,10 @@
   urldate = {2021-03-10}
 }
 
-@misc{matlab_2020_,
-  title = {Version: 9.9.0.1467703 ({{R2020b}})},
+@misc{matlab_2022_,
+  title = {Version: 9.13.0.2126072 ({{R2022b}})},
   author = {MATLAB},
-  year = {2020},
+  year = {2022},
   address = {{Natick, Massachusetts}},
   howpublished = {The MathWorks Inc.}
 }
