### 2023-09-07 c6fd6ba
docs: Rebuild all docs, update Contents
Lint all files: modern indenting, function end statements

Convert bfra_gettingStarted to GettingStarted.mlx in toolbox top-level

Add demo 3 to docs
