+++ +bfra/stationlist.m
@@ -1,9 +1,10 @@
 function stations = stationlist
-%STATIONLIST return list of stations from the bfra basin database
-% 
-% 
-% See also stationname
+   %STATIONLIST Return list of stations from the bfra basin database.
+   %
+   %
+   % See also: stationname
 
-% load(fullfile(fileparts([mfilename '.m']),'stationlist.mat'),'stations');
-load 'stationlist.mat' 'stations';
-stations = ['ALL_BASINS'; stations];
+   % load(fullfile(fileparts([mfilename '.m']),'stationlist.mat'),'stations');
+   load 'stationlist.mat' 'stations';
+   stations = ['ALL_BASINS'; stations];
+end
