+++ +bfra/cloudphi.m
@@ -1,140 +1,143 @@
 function [phi,a] = cloudphi(q,dqdt,blate,A,D,L,method,varargin)
-%CLOUDPHI estimate drainable porosity phi from the point cloud
-% 
-% Syntax
-% 
-%     [phi,a] = cloudphi(q,dqdt,blate,A,D,L,method,varargin)
-% 
-% Description
-% 
-%     [phi,a] = cloudphi(q,dqdt,blate,A,D,L,method) computes drainable porosity
-%     from discharge q, first derivative dqdt, aquifer properties area A, depth
-%     D, and channel length L, and late-time b-value blate for the event-scale
-%     recession equation -dq/dt = aQ^b, using the method of Troch, Troch, and
-%     Brutsaert, 1993.
-% 
-% Required inputs
-% 
-%     q           discharge (L T^-1, e.g. m d-1 or m^3 d-1)
-%     dqdt        discharge rate of change (L T^-2)
-%     blate       late-time b parameter in -dqdt = aq^b (dimensionless)
-%     A           basin area contributing to baseflow (L^2)
-%     D           saturated aquifer thickness (L)
-%     L           active stream length (L)
-%     method      method for fitting straight line to point cloud. Valid
-%                 options include 'median','mean','envelope'.
-% 
-% Optional name-value inputs
-% 
-%     refqtls     reference quantiles that together define a pivot point
-%                 through which the straight line must pass. use with method
-%                 'envelope'.
-%     mask        logical mask to exclude data
-%     theta       effective slope of basin contributing area
-%     isflat      logical flag indicating if horizontal or sloped aquifer
-%                 solution is applicable
-%     soln1       optional early-time theoretical solution
-%     soln2       optional late-time theoretical solution
-%     dispfit     logical flag indicating whether to plot the result
-% 
-% See also eventphi, fitphi, fitdistphi
-% 
-% Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
+   %CLOUDPHI Estimate drainable porosity phi from the point cloud.
+   %
+   % Syntax
+   %
+   %     [phi,a] = cloudphi(q,dqdt,blate,A,D,L,method,varargin)
+   %
+   % Description
+   %
+   %     [phi,a] = cloudphi(q,dqdt,blate,A,D,L,method) computes drainable
+   %     porosity from discharge q, first derivative dqdt, aquifer properties
+   %     area A, depth D, and channel length L, and late-time b-value blate for
+   %     the event-scale recession equation -dq/dt = aQ^b, using the method of
+   %     Troch, Troch, and Brutsaert, 1993.
+   %
+   % Required inputs
+   %
+   %     q           discharge (L T^-1, e.g. m d-1 or m^3 d-1)
+   %     dqdt        discharge rate of change (L T^-2)
+   %     blate       late-time b parameter in -dqdt = aq^b (dimensionless)
+   %     A           basin area contributing to baseflow (L^2)
+   %     D           saturated aquifer thickness (L)
+   %     L           active stream length (L)
+   %     method      method for fitting straight line to point cloud. Valid
+   %                 options include 'median','mean','envelope'.
+   %
+   % Optional name-value inputs
+   %
+   %     refqtls     reference quantiles that together define a pivot point
[…]
