+++ +bfra/private/isminlength.m
@@ -1,4 +1,5 @@
 function [tf, istart, iend] = isminlength(tf, nmin)
+   %ISMINLENGTH Return true for runlengths exceeding minimum length.
    rl = runlength(tf);                    % get run lengths
    tf = rl >= nmin;                       % find run lengths >= min length
    istart = find(diff([false; tf]) == 1); % start index of useable events
