+++ +bfra/private/isnumericscalar.m
@@ -1,3 +1,4 @@
 function tf = isnumericscalar(x)
+   %ISNUMERICSCALAR Return true if input X is a numeric scalar.
    tf = isnumeric(x) && isscalar(x);
 end

[…]
