### 2023-09-07 cb78e8a
refactor: move user-facing content to toolbox directory
move +test suite out of toolbox to top-level tests/ folder
