### 2023-09-07 fd93e06
fix: update paths in run-tests.yml
