### 2023-09-07 c1d1260
Squashed commit of the following:
commit 5f418e12697955a497bbcb9edf342698343313f8
Merge: 1bbe3d5 fd93e06
Author: mgcooper <mguycooper@gmail.com>
Date:   Wed Sep 6 19:16:02 2023 -0700

    Merge branch 'joss' into main-stage

commit fd93e0612bf66bb28cc589e5673529d45604da8e
Author: mgcooper <mguycooper@gmail.com>
Date:   Wed Sep 6 19:13:27 2023 -0700

    fix: update paths in run-tests.yml

commit cb78e8a669c8bf4c6a8246075caa3a1186358f05
Author: mgcooper <mguycooper@gmail.com>
Date:   Wed Sep 6 18:58:44 2023 -0700

    refactor: move user-facing content to toolbox directory

    move +test suite
[…]
