+++ toolbox/+bfra/dndtuncertainty.m
@@ -1,0 +1,283 @@
+function [sig_dndt, sig_lamda] = dndtuncertainty(T, Qb, Results, Fits, ...
+      GlobalFit, opts, varargin)
+   %DNDTUNCERTAINTY Compute combined uncertainty of the dn/dt trend estimate.
+   %
+   % Syntax
+   %
+   %     [sig_dndt, sig_lamda] = dndtuncertainty(T,Qb,Results,Fits,GlobalFit,opts)
+   %
+   % Description
+   %
+   %     [sig_dndt,sig_lamda] = dndtuncertainty(T,Qb,Results,Fits,GlobalFit,opts)
+   %     Computes the combined uncertainty (with correlation) for: dn/dt = 
+   %     lambda*dq/dt where dn/dt is the 'long term' (interannual) trend in
+   %     groundwater layer thickness (n), q is the 'long term' trend in
+   %     baseflow, and lambda = tau/phi*1/(N+1) is the linearized sensitivity
+   %     coefficient with reference drainage timescale tau [days], drainable
+   %     porosity [-], and exponent N [-], where N = 3-2b for all known
+   %     flat-aquifer and all known linearized sloped-aquifer solutions to the
+   %     one-dimensional groundwater flow equation for a Boussinesq aquifer, and
+   %     parameter b from -dQ/dt = aQb.
+   %
+   % See also: alttrend, aquiferthickness
+   %
+   % Matt Cooper, 04-Nov-2022, https://github.com/mgcooper
+
+   % if called with no input, open this file
+   if nargin == 0; open(mfilename('fullpath')); return; end
+
+   % parse inputs
+   alpha = 0.05;
+   testflag = false;
+   if nargin == 7
+      alpha = varargin{1};
+   elseif nargin == 8
+      alpha = varargin{1};
+      testflag = varargin{2};
+   end
+
+   % Convert time in days to years
+   if ~isdatetime(T)
+      T = datetime(T, 'ConvertFrom', 'datenum'); 
+   end
+   
+   [~, T] = padtimeseries(nan(size(T)), T, datenum(year(T(1)), 1, 1), ...
+      datenum(year(T(end)), 12, 31), 1); %#ok<*DATNM> 
+   T = transpose(year(mean(reshape(T, 365, numel(T)/365))));
+
+   if numel(T) ~= numel(Qb)
+      error('size T must match size Qb')
+   end
+
+   % Define the sensitivity coefficient and dn/dt trend functions
+   Flam = @(tau,phi,b) tau./(phi.*(4-2.*b));
+   Fdndt = @(tau,phi,b,dqdt) tau./(phi.*(4-2.*b)).*dqdt;
+
+   % Regress baseflow in units cm/day/year to get uncertainty on dq/dt using
+   % fitlm on the baseflow timeseries:
+   % Qb = GlobalFit.Qb./365.25;              % cm/yr -> cm/day
+   Qb = Qb./365.25;                          % cm/yr -> cm/day
+   mdl = fitlm(T,Qb); % plot(mdl)
+   dbfdt = mdl.Coefficients.Estimate(2);     % cm/day/year
+   std_dbfdt = mdl.Coefficients.SE(2);       % standard error
+   CI_dbfdt = mdl.coefCI;                    % 95% CI's
+   CI_dbfdt = CI_dbfdt(2,:);
+   sig_dbfdt = CI_dbfdt(2)-dbfdt;            % they're symetric so just take one
+
[…]
