+++ toolbox/+bfra/+deps/plfit.m
@@ -1,0 +1,310 @@
+function [alpha,xmin,L,D,n,dat]=plfit(x, varargin)
+%PLFIT Fit a power-law distributional model to data.
+%    
+%    Source: http://www.santafe.edu/~aaronc/powerlaws/
+%
+%    PLFIT(x) estimates x_min and alpha according to the goodness-of-fit
+%    based method described in Clauset, Shalizi, Newman (2007). x is a
+%    vector of observations of some quantity to which we wish to fit the
+%    power-law distribution p(x) ~ x^-alpha for x >= xmin.
+%    PLFIT automatically detects whether x is composed of real or integer
+%    values, and applies the appropriate method. For discrete data, if
+%    min(x) > 1000, PLFIT uses the continuous approximation, which is
+%    a reliable in this regime.
+%
+%    The fitting procedure works as follows:
+%    1) For each possible choice of x_min, we estimate alpha via the
+%       method of maximum likelihood, and calculate the Kolmogorov-Smirnov
+%       goodness-of-fit statistic D.
+%    2) We then select as our estimate of x_min, the value that gives the
+%       minimum value D over all values of x_min.
+%
+%    Note that this procedure gives no estimate of the uncertainty of the
+%    fitted parameters, nor of the validity of the fit.
+%
+%    Example:
+%       x = (1-rand(10000,1)).^(-1/(2.5-1));
+%       [alpha, xmin, L] = plfit(x);
+%
+%    The output 'alpha' is the maximum likelihood estimate of the scaling
+%    exponent, 'xmin' is the estimate of the lower bound of the power-law
+%    behavior, and L is the log-likelihood of the data x>=xmin under the
+%    fitted power law.
+%
+%    For more information, try 'type plfit'
+%
+%    See also PLVAR, PLPVA
+
+% Version 1.0    (2007 May)
+% Version 1.0.2  (2007 September)
+% Version 1.0.3  (2007 September)
+% Version 1.0.4  (2008 January)
+% Version 1.0.5  (2008 March)
+% Version 1.0.6  (2008 July)
+% Version 1.0.7  (2008 October)
+% Version 1.0.8  (2009 February)
+% Version 1.0.9  (2009 October)
+% Version 1.0.10 (2010 January)
+% Version 1.0.11 (2012 January)
+% 
+% LICENSE
+% 
+% Copyright (C) 2008-2012 Aaron Clauset (Santa Fe Institute)
+% Distributed under GPL 2.0
+% http://www.gnu.org/copyleft/gpl.html
+% PLFIT comes with ABSOLUTELY NO WARRANTY
+%
+% Notes:
+%
+% 1. In order to implement the integer-based methods in Matlab, the numeric
+%    maximization of the log-likelihood function was used. This requires
+%    that we specify the range of scaling parameters considered. We set
+%    this range to be [1.50 : 0.01 : 3.50] by default. This vector can be
+%    set by the user like so,
+%
+%       a = plfit(x,'range',[1.001:0.001:5.001]);
+%
+% 2. PLFIT can be told to limit the range of values considered as estimates
+%    for xmin in three ways. First, it can be instructed to sample these
+%    possible values like so,
+%
+%       a = plfit(x,'sample',100);
+%
+%    which uses 100 uniformly distributed values on the sorted list of
+%    unique values in the data set. Second, it can simply omit all
+%    candidates above a hard limit, like so
+%
+%       a = plfit(x,'limit',3.4);
+%
+%    Finally, it can be forced to use a fixed value, like so
+%
+%       a = plfit(x,'xmin',3.4);
+%
[…]
