+++ +bfra/private/isdatelike.m
@@ -1,3 +1,0 @@
-function tf = isdatelike(x)
-   tf = isdatetime(x) || (isnumeric(x) && isvector(x));
-end

[…]
