### 2023-09-07 #5492 @mgcooper
@elbeejay @tianydong @alessandroamaranto @hydrotian

Revisions are complete and pushed to the joss branch (also main). Updates include:

- All tests pass.
- All demo files run as both live scripts in .mlx format and m-files.
- The docs were recompiled.
- The paper was recompiled.

Some important changes were made to the repository structure:

- The paper files were moved from the top-level to the new `paper/` folder.
- User-facing "toolbox" contents were moved from the top-level to the new `toolbox/` folder.
- Developer-facing contents were retained in the top-level.

From a user-perspective, the reorganization only affects the tests. Instead of typing `bfra.test.runtests`, just type `runtests`. Overall, the new structure aligns with standard software development practices. I think it is important for the release associated with this paper.

Please see a point by point reply to reviewer comments below.

@tianydong "Overall the package is easy to install, the codes are well-commented, and demo 1 runs fine. But I am getting some errors on the other demos, for example: <datetime related error>"
- This was addressed by adding a new function `todatenum` that ensures datetimes are converted to datenums. See commit 98499ac91f2987678c63ce1bb74c4c9b78662475.

@tianydong "Maybe a path thing? Also, I am using MATLAB 2022b, but I wonder if that's a problem."
- This should not be a problem. I have also upgraded to 2022b and everything runs fine. Just FYI.

@tianydong "Also, there's a conflict between the .m and .mlx files for the linear theory demos."
- This was addressed by moving the m-file demos to the `demos/mfiles` sub-folder so they don't shadow the live script .mlx files in the `demos` folder. Users with access to Matlab are encouraged to use the .mlx files, whereas Octave users can use the m-files, but the content is identical and both live scripts and m-files run in Matlab. See commit 0b9b00cee415bfececef26a2edcad53961da31a9.

We believe this satisfies all requests from reviewers. Please let us know if any issues are encountered running the software. Many thanks for the review.

Matt Cooper
Tian Zhou
