### 2023-09-12 #5492 @mgcooper
@tianydong  We did update the paper, our apologies for omitting that from the response. Specifically, to address each of your comments:

1. "maybe add one or two examples of aquifer properties, like K or porosity, here"
As suggested, we added examples of aquifer properties to the abstract.

2. "I'd suggest adding one sentence here to quickly define baseflow."
As suggested, we added a sentence to the first paragraph to define baseflow. We also broke the first paragraph into two. The first paragraph includes the baseflow definition, as requested, and a statement about why baseflow matters. The second paragraph addresses comment #3, specifically "how exactly recession curve can be used to infer ... aquifer properties".

3. "If space permit, I'd also suggest elaborate how exactly recession curve can be used to infer geological structures, particularly to the aquifer properties that baseflow is able to infer (mentioned later in the paper)"

We removed the phrase "geologic structures" to avoid confusion among those with a groundwater hydrology background. We also removed this phrase from the abstract. To answer your question, the geologic structures themselves cannot be inferred, rather the theory is based on an assumption that the aquifer is unconfined. To address this, we emphasize "unconfined aquifer properties" throughout the paper, and removed "geologic structures".

To address the other side of your comment: "how exactly recession curve can be used to infer ... aquifer properties", see our first two answers above. Additionally, we added a sentence to the first paragraph of the "State of the field" section that directly addresses this question: "Several well-known solutions to the one-dimensional lateral groundwater flow equation for unconfined aquifers can be written in the same form as [the recession equation]" and this forms the basis for estimating aquifer properties from the parameters obtained by curve-fitting recession sequences.

We also added a sentence immediately before "State of the field" that mentions the two theory-focused notebooks, which also provide more specific information about how recession curves can be used to infer aquifer properties.

4. "I'd suggest explain each panel in the figure caption"
As requested, we added an explanation of each panel to the figure caption.

Finally, we made various edits to improve the writing quality and/or grammar. Please see the compiled pdf in the paper/ folder at the top-level of the repo.

Thank you,

Matt Cooper
Tian Zhou
