### 2023-09-12 3bed506
refactor(+internal): Move internal toolbox functions to +internal.
Move all makedocs related functions to +internal.

Move basepath() and any path-related functions to +internal.

Replace +bfra prefix with +baseflow in related function calls.

Replace +bfra prefixes with +baseflow in functionSignatures.

Replace +bfra prefixes with +baseflow in helptoc.xml and GettingStarted.mlx

Preliminary prefix replacements before running programmatic replacements.
