+++ toolbox/+baseflow/+internal/demopath.m
@@ -1,4 +1,4 @@
 function fullpath = demopath(varargin)
    %DEMOPATH Return toolbox demo path.
-   fullpath = fullfile(basepath(), 'demos', varargin{:});
+   fullpath = fullfile(baseflow.internal.basepath(), 'demos', varargin{:});
 end
