### 2023-09-12 1c9ce88
refactor(+baseflow): Replace all +bfra prefixed function calls with +baseflow.
Recompile and publish demos.

Rebuild docsearchdb.
