+++ toolbox/docs/baseflow_examples_contents.m
@@ -2,20 +2,20 @@
 %  
 % These examples provide an introduction to the toolbox. The purpose of these
 % examples is to demonstrate a typical use-case for the main functions in the
-% bfra toolbox. To explore the toolbox functions in more detail, see the
+% baseflow toolbox. To explore the toolbox functions in more detail, see the
 % <m2html/function_index.html Function Index>.
 % 
-% * <bfra_demo_1.html *Estimate recession constants: Point cloud*> shows how
+% * <baseflow_demo_1.html *Estimate recession constants: Point cloud*> shows how
 % to detect recession events from a timeseries of observed streamflow discharge
 % and estimate recession constants using the point cloud method. 
-% * <bfra_demo_2.html *Estimate recession constants: Pareto distribution*>
+% * <baseflow_demo_2.html *Estimate recession constants: Pareto distribution*>
 % shows how to detect recession events from a timeseries of observed streamflow
 % discharge and estimate recession constants using a hybrid point-cloud and
 % Pareto distribution fitting method.
-% * <bfra_demo_3.html *Aquifer property estimation: Introduction*>
+% * <baseflow_demo_3.html *Aquifer property estimation: Introduction*>
 % shows how to estimate aquifer properties from curve fits to baseflow recession
 % events.
-% * <bfra_demo_kuparuk.html *Estimate saturated aquifer thickness trends using
+% * <baseflow_demo_kuparuk.html *Estimate saturated aquifer thickness trends using
 % baseflow recession analysis*> shows how to detect recession events from a
 % timeseries of observed streamflow discharge and estimate recession constants
 % using a hybrid point-cloud and Pareto distribution fitting method, and then
