+++ toolbox/+baseflow/Contents.m
@@ -1,185 +1,188 @@
-% +BFRA
+% +BASEFLOW
 % 
-%   Contents file for +BFRA and its subfolders.
+%   Contents file for +BASEFLOW and its subfolders.
 %   
-%   +BFRA
-%   baseflow.aQbString                 - Return a formatted string for equation aQ^b
-%   baseflow.aquiferprops              - Estimate aquifer depth, hydraulic conductivity, and porosity
-%   baseflow.aquiferstorage            - Estimate aquifer storage from recession event parameters
-%   baseflow.aquiferthickness          - Estimate aquifer thickness from recession event parameters
-%   baseflow.aquifertrend              - Estimate the linear trend in saturated aquifer thickness
-%   baseflow.baseflowtrend             - Estimate baseflow trend from annual streamflow timeseries
-%   baseflow.basinlist                 - Load the list of basins in the baseflow database
-%   baseflow.basinname                 - Return string 'basin' from the baseflow basin database
-%   baseflow.characteristicTime        - Compute the characteristic e-folding time for baseflow
-%   baseflow.checkevent                - Plot detected recession event and fitted values
-%   baseflow.cloudphi                  - Estimate drainable porosity phi from the point cloud
-%   baseflow.completions               - Generate function auto-completions for string literals
-%   baseflow.conversions               - Convert inputvalue of inputvarname to the value of outputvarname
-%   baseflow.dependencies              - Generate function and product dependencies for function
-%   baseflow.dndtuncertainty           - Compute combined uncertainty of the dn/dt trend estimate
-%   baseflow.eventfinder               - Find recession events using hydrograph and rainfall timeseries
-%   baseflow.eventphi                  - Estimate drainable porosity phi from individual recession events
-%   baseflow.eventpicker               - Automated or user-guided recession event selection
-%   baseflow.eventplotter              - Plot recession events detected by eventfinder
-%   baseflow.eventtau                  - Compute drainage timescale tau from event-scale parameters a and b
-%   baseflow.expectedQ                 - Compute the expected value of baseflow
-%   baseflow.fdcurve                   - Compute a flow duration curve from streamflow timeseries
-%   baseflow.fitab                     - Fit event-scale recession equation -dq/dt = aQ^b
-%   baseflow.fitevents                 - Wrapper around getdqdt and fitdqdt functions to fit all events
-%   baseflow.fitphi                    - Estimate drainable porosity using early- and late-time solutions
-%   baseflow.fitphidist                - Fit drainable porosity (phi) values with a Beta distribution
-%   baseflow.generateTestData          - Generate test data for baseflow recession analysis
-%   baseflow.getdqdt                   - Numerical estimation of the time derivative of discharge dQ/dt
-%   baseflow.getevents                 - Get recession events from daily hydrograph timeseries T, Q, and R
-%   baseflow.getEventsData             - Get data from Events for event number eventTag
-%   baseflow.getFitsData               - Get data from Fits for event number eventTag
-%   baseflow.getfunction               - Get function handle from the baseflow function library
-%   baseflow.getstring                 - Get latex-formatted string for equations in the baseflow library
-%   baseflow.globalfit                 - Fit global parameters using all event-scale recession fits
-%   baseflow.gpfitb                    - Fit Generalized Pareto Distribution to recession parameter tau
-%   baseflow.help                      - Open toolbox html help document in the MATLAB Help browser
-%   baseflow.hyetograph                - Plot a discharge rainfall hyetograph
-%   baseflow.loadbasins                - Load boundary object for basin specified by basinname
-%   baseflow.loadcalm                  - Load CALM ALT data for a basin in the Bounds struct
-%   baseflow.loadExampleData           - Load toolbox example data
[…]
