+++ toolbox/+baseflow/+util/Contents.m
@@ -7,4 +7,4 @@
 %   baseflow.util.numfits   - Count the number of fits in struct returned by baseflow.fitevents
 %   baseflow.util.numtau    - Count the number of tau values in struct returned by baseflow.fitevents
 %    
-%   This file was generated by updatecontents.m on 06 Sep 2023 at 15:36:12.
+%   This file was generated by updatecontents.m on 12 Sep 2023 at 00:41:17.
