+++ toolbox/+baseflow/+internal/Contents.m
@@ -1,0 +1,12 @@
+% +INTERNAL
+%
+% Files
+%   baseflow.internal.basepath               - Return toolbox basepath
+%   baseflow.internal.completions            - Generate function auto-completions for string literals
+%   baseflow.internal.convertlivescripts     - Convert live script to m-files
+%   baseflow.internal.demopath               - Return toolbox demo path
+%   baseflow.internal.dependencies           - Generate function and product dependencies for function
+%   baseflow.internal.docpath                - Return path to toolbox doc file
+%   baseflow.internal.makedocs               - Publish GettingStarted.m as index.html for GitPages
+%   baseflow.internal.runtests               - Run all tests in the baseflow test suite
+%   baseflow.internal.version                - Return the version number for the baseflow toolbox
