### 2023-09-19 #5492 @elbeejay
@mgcooper if you could notify me when you've done the author items on the above list that'd be great (as it's my comment so you won't be able to edit that checklist directly). Then I'll go ahead and go through the editor part of the checklist and we'll be ready to recommend this paper for publication. Thanks!
