### 2023-09-20 2c1ec9e
refactor(+internal): Generalize +internal by moving all function dependencies to private/.
Make +internal functions namespace-independent, and dependent only on built-in or private/ functions.

Replace calls to internal.basepath() with internal.buildpath and/or projectpath() and toolboxpath().

Add internal function test suite.

Add makecontents to +internal.
