+++ toolbox/+baseflow/+internal/private/projectpath.m
@@ -1,0 +1,5 @@
+function fullpath = projectpath()
+   %PROJECTPATH Return project basepath.
+   fullpath = fileparts(fileparts(fileparts(fileparts(fileparts( ...
+      mfilename('fullpath'))))));
+end
