### 2023-09-20 c62858a
fix(octave-compat): Generate default opts struct in fitets using optimset.
Add pkg load financial for date functions in wrapfits.

Add private folders to path.

Patch convertStringsToChars.

Update .octaverc.

Add isoctave check before addListener in hyetograph.

The addlistener to update yyaxis ylabel is not octave compatible.

baseflow.open: Use edit with full path.
