### 2023-09-30 #5492 @elbeejay
@mgcooper what's your estimated timeline for taking care of these last few items? They should be pretty quick, they exist to ensure the metadata associated with the JOSS publication is up to date.
