+++ toolbox/+baseflow/+internal/dependencies.m
@@ -11,13 +11,12 @@ function report = dependencies(funcname,option)
    %
    % See also: Setup
 
-   % use this to generate a list of all functions in the package, then cycle over
-   % all of them and find the dependencies
-   % funcpath = fileparts(which('baseflow.fitab'));
+   % use this to generate a list of all functions in the package, then cycle
+   % over all of them and find the dependencies
+   % funcpath = fileparts(which('pkg.func'));
    % funclist = getlist(funcpath,'.m');
 
-   % if called with no input, open this file
-   if nargin == 0; open(mfilename('fullpath')); return; end
+   [pkgname, pkgfolder] = mpackagename();
 
    % valid options
    validopts = ["all","missing","installed","resolve","report","check"];
@@ -26,16 +25,13 @@ function report = dependencies(funcname,option)
    end
 
    % use this to test a particular function
-   % if nargin == 0
-   % funcname = 'baseflow_demo.mlx';
-   % funcname = 'baseflow_gettingStarted.m';
-   % funcname = 'baseflow.fitab';
-   % funcname = 'baseflow_kuparuk.m';
-   % end
+   if nargin == 0
+      funcname = '';
+   end
 
    % this loads the saved dependencies.mat file and checks against util/
    if opts.check == true
-      report = dependencycheck();
+      report = dependencycheck(pkgfolder);
       return;
    end
 
@@ -48,9 +44,9 @@ function report = dependencies(funcname,option)
       % return a table of all dependencies
       report = cell2table(funclist,'VariableNames',{'function_dependencies'});
    elseif opts.report == true
-      report = getdependencyreport(funclist,prodlist,funcname);
+      report = getdependencyreport(funclist,prodlist,funcname,pkgfolder);
    elseif opts.missing == true
-      report = getmissingdependencies(funclist,funcname);
+      report = getmissingdependencies(funclist,funcname,pkgname,pkgfolder);
    elseif opts.installed == true
       report = getinstalleddependencies(funclist);
    elseif opts.resolve == true
@@ -58,12 +54,12 @@ function report = dependencies(funcname,option)
    end
 end
 
-function report = getdependencyreport(funclist,prodlist,funcname)
+function report = getdependencyreport(funclist,prodlist,funcname,pkgfolder)
 
    % get a list of dependencies that are not toolbox functions, i.e. those that
    % need to be included in util/
    % report = funclist;
-   skip = {'+baseflow',funcname,'ExtractNameVal','Cupid'};
+   skip = {pkgfolder, funcname, 'ExtractNameVal', 'Cupid'};
    keep = true(numel(funclist),1);
    for n = 1:numel(funclist)
       keep(n) = ~ismember(funclist{n},skip);
@@ -73,8 +69,8 @@ function report = getdependencyreport(funclist,prodlist,funcname)
    % report = report(keep);
 end
 
-function report = dependencycheck
-   load(fullfile(getinstallpath(), '+baseflow', 'private', 'dependencies.mat'), 'report')
+function report = dependencycheck(pkgfolder)
+   load(fullfile(getinstallpath(), pkgfolder, 'private', 'dependencies.mat'), 'report')
    deps = report.function_dependencies;
    deps = deps(~isfile(fullfile(getinstallpath(), 'util', deps)));
    deps = deps(~isfolder(strrep(fullfile(getinstallpath(), 'util', deps),'.m','')));
@@ -86,15 +82,15 @@ function report = dependencycheck
    end
 end
 
-function report = getmissingdependencies(funclist,funcname)
+function report = getmissingdependencies(funclist,funcname,pkgname,pkgfolder)
 
-   % remove functions in the toolbox first, leave the ones in util/ so they can be
-   % compared here if needed for debugging.
+   % remove functions in the toolbox first, leave the ones in util/ so they can
[…]
