+++ toolbox/+baseflow/+internal/private/isoctave.m
@@ -1,0 +1,13 @@
+function tf = isoctave()
+   %ISOCTAVE return true if the environment is Octave.
+   %
+   %
+   % See also
+
+   persistent cacheval
+
+   if isempty (cacheval)
+      cacheval = (exist ("OCTAVE_VERSION", "builtin") > 0);
+   end
+   tf = cacheval;
+end
