### 2023-10-02 88ce3a0
feat(fitlm_octmat): Add Octave-compatible fitlm, predidtlm, and smooth.
Add pkg load statistics-bootstrap.

Remove PD from fitphidist.

Refactor baseflowtrend, plotaquifertrend, trendplot for Octave.

Replace PD object in phi fitting with beta* functions.
