+++ toolbox/+baseflow/fitphidist.m
@@ -13,8 +13,8 @@ function [Fit,h] = fitphidist(phi,varargin)
    %     'Fit' which is a Beta Distribution fit to the input data in phi
    %
    %     Fit = baseflow.fitphidist(phi,outputtype) returns a Beta Distribution fit
-   %     to the input data in phi. 'outputtype' is 'PD', 'mean', 'median', or
-   %     'std', where default 'PD' is the Probability Distribution object.
+   %     to the input data in phi. 'outputtype' is 'PD', 'mean', or 'std', where
+   %     default 'PD' is the Probability Distribution object. 
    %
    %     Fit = baseflow.fitphidist(__,plottype) returns any of the prior options
    %     plus a figure showing the fit. plottype can be 'cdf' or 'pdf'. default
@@ -34,21 +34,28 @@ function [Fit,h] = fitphidist(phi,varargin)
    phi = phi(:);
 
    % Fit the beta distribution
-   PD = fitdist(phi, 'beta');
+   [PARAMHAT, PARAMCI] = betafit(phi);
+   [M, V] = betastat(PARAMHAT(1), PARAMHAT(2));
+   PD.paramEst = PARAMHAT;
+   PD.paramCI = PARAMCI;
+   PD.a = PARAMHAT(1);
+   PD.b = PARAMHAT(2);
+   PD.mean = M;
+   PD.std = sqrt(V);
 
    % Create the figure
    switch plottype
       case 'cdf'
-         h = cdfplotphi(phi,PD,showfit);
+         h = cdfplotphi(phi, PD, showfit);
       case 'probplot'
-         h = probplotphi(phi,PD,showfit);
+         h = probplotphi(phi, PD, showfit);
       case 'pdf'
    end
 
    % package output
    switch outputtype
       case 'PD'
-         % send back the ProbabilityDistribution object
+         % send back the probability distribution
          Fit = PD;
       case 'mean'
          % send back the mean
@@ -56,82 +63,60 @@ function [Fit,h] = fitphidist(phi,varargin)
       case 'std'
          % send back the mean
          Fit = PD.std;
-      case 'median'
-         Fit = PD.median;
    end
 end
 
-function h = cdfplotphi(phi,PD,showfit)
-
-   % Create the figure
-   h.figure = figure('Visible',showfit);
+function h = cdfplotphi(phi, PD, showfit)
 
    % Get the cdf
-   [cdfY,cdfX] = ecdf(phi,'Function','cdf');  % compute empirical function
-   % h.data = stairs(cdfX,cdfY,'LineWidth',1); hold on;
-
-   c = [0 0.447 0.741];
-
-   h.data = plot(cdfX,cdfY,'o','MarkerFaceColor',c,'MarkerEdgeColor','none', ...
-      'Visible',showfit); hold on;
+   [cdfY, cdfX] = ecdf(phi, 'Function', 'cdf');
+   
+   % Create the figure
+   h.figure = figure('Visible', showfit);
+   h.data = plot(cdfX, cdfY, 'o', 'MarkerFaceColor', [0 0.447 0.741], ...
+      'MarkerEdgeColor', 'none', 'Visible', showfit); hold on;
 
    % Create grid where function will be computed
-   XLims = get(gca,'XLim');
+   XLims = get(gca, 'XLim');
    XLims = XLims + [-1 1] * 0.01 * diff(XLims);
-   XGrid = linspace(XLims(1),XLims(2),100);
[…]
