### 2023-10-02 f7a5ef6
refactor(+internal/mpackagename): Replace hard-coded namespace with auto-detected name.
