+++ toolbox/demos/mfiles/baseflow_demo_kuparuk.m
@@ -1,18 +1,18 @@
 %% Estimate active layer thickness trends using baseflow recession analysis
-% This script runs the analysis for the Kuparuk River Basin reported in Cooper
-% et al. 2023, <https://doi.org/10.1029/2022WR033154 Detecting Permafrost Active
-% Layer Thickness Change From Nonlinear Baseflow Recession>.
-%
+% This script runs the analysis for the Kuparuk River Basin reported in Cooper 
+% et al. 2023, <https://doi.org/10.1029/2022WR033154 Detecting Permafrost Active 
+% Layer Thickness Change From Nonlinear Baseflow Recession>. 
+% 
 % Notes for users:
-%%
-% * This demo is designed to replicate the experiment noted above, therefore
-% it may use different notation, terminology, or other conventions than the current
+%% 
+% * This demo is designed to replicate the experiment noted above, therefore 
+% it may use different notation, terminology, or other conventions than the current 
 % toolbox version.
 % * Detecting recession events is reasonably fast but fitting events is slow.
 % * Computing uncertainties is very slow because it uses a bootstrap fit.
-% * Therefore, an option is included to either run each step or load pre-computed
+% * Therefore, an option is included to either run each step or load pre-computed 
 % data.
-% * In typical use, one might run this script once to generate the 'Events'
+% * In typical use, one might run this script once to generate the 'Events' 
 % data and save that.
 % * Then experiment with the 'fitevents' function and save the 'Fits' data.
 % * Then experiment with the 'globalfit' function and save the 'GlobalFit' data.
@@ -20,7 +20,7 @@
 clearvars
 close all
 clc
-%%
+%% 
 % Set the main options
 
 savedata    = false;
@@ -33,23 +33,23 @@ plotfigs    = true; % if true, globalfit will produce figures
 bootfit     = false; % if true, bootstrap uncertainties (very slow)
 nreps       = 1000; % number of bootstrap samples
 
-%%
+%% 
 % Set the filename to save the Events, Fits, and GlobalFit .mat files.
 
 fname = fullfile(baseflow.internal.buildpath('data'), 'Events.mat');
-%%
+%% 
 % Set the basin geomorphological parameters
 
 A  = 8.6545e9;       % basin area
 Dd = 0.8;            % drainage density
 D  = 0.47;           % reference aquifer thickness (mean alt from CALM)
 L  = A*Dd/1000;      % active stream length
-%%
+%% 
 % Load the streamflow data and the CALM-network active layer thickness data
 
 [T, Q, R, Data] = baseflow.loadExampleData('kuparuk');
-%%
-% Set the algorithm options. Here we use a combination of default options and
+%% 
+% Set the algorithm options. Here we use a combination of default options and 
 % site-specific options.
 
 opts.getevents = baseflow.setopts('getevents', 'asannual', true);
@@ -58,7 +58,7 @@ opts.globalfit = baseflow.setopts('globalfit', ...
    'drainagearea', A, 'aquiferdepth', D, ...
    'streamlength', L, 'drainagedensity', Dd, ...
    'bootfit', bootfit, 'bootreps', nreps, 'plotfits', true);
-%%
+%% 
 % Get all recession events and then fit them
 
 if fitevents == true
@@ -72,8 +72,8 @@ if fitevents == true
 else % load pre-saved data
[…]
