### 2023-10-03 f426150
chore(linting/data): Apply linting to deps.
Upload octave compatible data.
