+++ toolbox/+baseflow/+deps/movingslope.m
@@ -1,158 +1,239 @@
 function Dvec = movingslope(vec,supportlength,modelorder,dt)
-%MOVINGSLOPE Estimate local slope from point sequence using a sliding window.
-% 
-% usage: Dvec = movingslope(vec)
-% usage: Dvec = movingslope(vec,supportlength)
-% usage: Dvec = movingslope(vec,supportlength,modelorder)
-% usage: Dvec = movingslope(vec,supportlength,modelorder,dt)
-%
-%
-% movingslope uses filter to determine the slope of a curve stored
-% as an equally (unit) spaced sequence of points. A patch is applied
-% at each end where filter will have problems. A non-unit spacing
-% can be supplied.
-%
-% Note that with a 3 point window and equally spaced data sequence,
-% this code should be similar to gradient. However, with wider
-% windows this tool will be more robust to noisy data sequences.
-%
-%
-% arguments: (input)
-%  vec - row of column vector, to be differentiated. vec must be of
-%        length at least 2.
-%
-%  supportlength - (OPTIONAL) scalar integer - defines the number of
-%        points used for the moving window. supportlength may be no
-%        more than the length of vec.
-%
-%        supportlength must be at least 2, but no more than length(vec)
-%
-%        If supportlength is an odd number, then the sliding window
-%        will be central. If it is an even number, then the window
-%        will be slid backwards by one element. Thus a 2 point window
-%        will result in a backwards differences used, except at the
-%        very first point, where a forward difference will be used.
-%
-%        DEFAULT: supportlength = 3
-%
-%  modelorder - (OPTIONAL) - scalar - Defines the order of the windowed
-%        model used to estimate the slope. When model order is 1, the
-%        model is a linear one. If modelorder is less than supportlength-1.
-%        then the sliding window will be a regression one. If modelorder
-%        is equal to supportlength-1, then the window will result in a
-%        sliding Lagrange interpolant.
-%
-%        modelorder must be at least 1, but not exceeding
-%        min(10,supportlength-1)
-%
-%        DEFAULT: modelorder = 1
-%
-%  dt - (OPTIONAL) - scalar - spacing for sequences which do not have
-%        a unit spacing.
-%
-%        DEFAULT: dt = 1
-%
-% arguments: (output)
-%  Dvec = vector of derivative estimates, Dvec will be of the same size
-%        and shape as is vec.
-% 
-%
-% Example:
-%  Estimate the first derivative using a 7 point window with first through
-%  fourth order models in the sliding window. Note that the higher order
-%  approximations provide better accuracy on this curve with no noise.
-%  
-%  t = 0:.1:1;
-%  vec = exp(t);
-%
-%  Dvec = movingslope(vec,7,1,.1)
-%  Dvec =
-%  Columns 1 through 7
-%    1.3657  1.3657  1.3657  1.3657  1.5093  1.668  1.8435
-%  Columns 8 through 11
-%    2.0373  2.0373  2.0373  2.0373
-%
-%  Dvec = movingslope(vec,7,2,.1)
-%  Dvec =
-%  Columns 1 through 7
-%    0.95747 1.0935  1.2296  1.3657  1.5093  1.668  1.8435
-%  Columns 8 through 11
-%    2.0373  2.2403  2.4433  2.6463
-%
-%  Dvec = movingslope(vec,7,3,.1)
-%  Dvec =
-%  Columns 1 through 7
-%    1.0027  1.1049  1.2206  1.3498  1.4918  1.6487  1.8221
[…]
