### 2023-10-03 78b833e
chore: Update version and doi in citation file.
