+++ CITATION.cff
@@ -14,8 +14,8 @@ authors:
     orcid: 'https://orcid.org/0000-0002-0165-209X'
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.7373924
-    description: The DOI for version 0.1.0 of the software
+    value: 10.5281/zenodo.8401281
+    description: The DOI for version 1.0.0 of the software
 repository-code: 'https://github.com/mgcooper/baseflow'
 url: 'https://github.com/mgcooper/baseflow'
 keywords:
@@ -25,5 +25,5 @@ keywords:
   - storage
   - permafrost
 license: BSD-3-Clause
-version: 0.1.0
-date-released: '2022-11-28'
+version: 1.0.0
+date-released: '2023-10-02'
