### 2023-10-03 2e651c0
chore: Update version and doi in citation file.
