### 2023-10-03 dfe07bb
chore: Update version and doi in citation file.
