### 2023-10-03 #5492 @mgcooper
@elbeejay

I completed the check list. Below are my replies to action items:

Release version is v1.0.0. I made a v1.0.0-joss-branch release which used the joss branch as the target and another on main branch. Feel free to use whichever you prefer, but I wanted a v1.0.0 release to point to main for future clarity.

[10.5281/zenodo.8401301](https://doi.org/10.5281/zenodo.8401301)

Please let me know if anything remains. Thank you,
Matt
