### 2023-10-03 #5492 @elbeejay
@mgcooper is there a reason the license is in a file called `COPYING`? as opposed to one called `LICENSE`? I think the latter would be our preference as it is more standard.
