### 2023-10-05 b19e56f
chore: Update title in zenodo file and other metadata to match JOSS paper.
