### 2023-10-05 #5492 @mgcooper
> Hi @mgcooper - I need the title of the zenodo entry to match the title of the paper, if you could do that it'd be great.

Hi @elbeejay Are you referring the "title" field of the .zenodo.json file? If so, I updated it as requested.

Regarding the license file, I changed it to LICENSE as requested. I am not sure why it was COPYING but apparently that is the GNU standard. I find it off-putting so I am glad that you asked to change it.

Please let me know what else I can do. Thank you,
Matt
