### 2023-10-07 #5492 @elbeejay
Hi @mgcooper, I am referring to the title as shown in the sceenshot below:

![image](https://github.com/openjournals/joss-reviews/assets/1770513/315f6921-c858-41dd-9614-9efec49adc36)

JOSS asks that the title and authors (and ORCID information etc.) shown there in Zenodo match what's on the paper.

Hope that helps clarify.

Thanks
