### 2023-10-16 #5492 @mgcooper
@elbeejay @kthyng  Sounds good, thanks. I did want to flag the latex formatting in the .md file as a possible source for the error. Also, since the error is related to the paper.jats file, I just wanted to mention that this file is generated when I compile the paper locally using my compile.paper script. I assume an existing .jats file alongside the .md file is not problematic for the JOSS software, but maybe there is something about the way it's generated by compile.paper that is problematic? The script is just a slightly modified version of one that (I think) is posted on a JOSS-affiliated repo, but I can't recall right now where I found it. I think I just added the redirect to /dev/null/ but I might have modified it slightly in some other way.
