### 2023-10-21 #5492 @elbeejay
Thanks for the description @mgcooper - I'll ping the wider group of JOSS editors to see if anyone has encountered this before.
