### 2023-10-23 #5492 @xuanxu
I think the error comes from the use of `dfrac` [in the paper.md](https://github.com/mgcooper/baseflow/blob/joss/paper/paper.md?plain=1#L55) file.
Currently setting a fraction in displaystyle using `\dfrac` is not totally supported by the JOSS compilation tool. Changing it to `\frac` should fix the problem.
