### 2023-10-23 6368e73
fix(paper): Change dfrac to frac for compatibility with joss compiler.
