### 2023-10-23 #5492 @mgcooper
@elbeejay @xuanxu

I changed to `\dfrac` to `\frac` and pushed an update. Please let me know if it fixes the issue. Thanks!
