### 2023-10-23 #5492 @mgcooper
@elbeejay

The proofs look good except one minor issue. If it is not too late, they recently updated the name of the division we are affiliated with at PNNL. I just pushed an update to the paper.md file to reflect this. If possible, please recompile the paper using that version. If not, no worries, it is not a big deal. Thanks!
Matt
