### 2023-10-24 #5492 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
